Crate sapphire

Sapphire

These are the basic APIs for building, manipulating and emitting SIR.

Modules

analysis

Contains the various analysis passes defined in the Sapphire project.

arena

A simple typed arena module.

cli

Contains utility code specifically for the CLI tools located in the tools/ subdirectory.

ir

Provides the interfaces and the types required to properly manipulate SIR.

passes

Defines the pass infrastructure used by the middle-end of the compiler.

utility

Provides several utility APIs that are used inside of various modules inside of the compiler.

Macros

arena_key

Creates a type-safe key for a ArenaMap and associated data structures.

dense_arena_key

Creates a type-safe key for a ArenaMap with u32 as the underlying data type. Acts just like arena_key but with u32 being the default type instead of the unspecified default.

Functions

parse_sir

Parses a SIR source file into a module, or returns the first error that occurred while parsing.