Module sapphire::analysis

Contains the various analysis passes defined in the Sapphire project.

These are basically all passes that model the FunctionAnalysisPass or the ModuleAnalysisPass traits, and range from debug passes to analyses that are critical for correctness.

Structs

CFGAnalysis

An analysis pass that wraps up a ControlFlowGraph into something that can actually be used inside of transform passes.

ControlFlowGraph

Models successor/predecessor information about the control-flow graph of a given function.

ModuleStringifyAnalysis

This is an analysis that provides a ModuleWriter to any code that wants it.

ModuleWriter

A simple SIR -> text pass that takes in an entire module, turns it into textual SIR, and then maps each IR entity to a range of text referring to it.

ModuleWriterAnalysis

This is an analysis that writes out a textual representation of a module to a

Functions

print_module

Prints an entire module to stdout.

stringify_module

Writes an entire module to a String.