Struct sapphire::analysis::ModuleWriter

pub struct ModuleWriter { /* private fields */ }

A simple SIR -> text pass that takes in an entire module, turns it into textual SIR, and then maps each IR entity to a range of text referring to it.

This can be used for debug/test passes that need to produce human-readable SIR.

Implementations

impl ModuleWriter

pub fn ty(&self, ctx: &TypePool, ty: Type) -> String

Stringifies a SIR type. Unlike the rest of this module, this type does not necessarily have to have been used in the module that this was constructed from.

pub fn val(&self, value: Value) -> &str

Provides the name of a value. This is the %_ syntax.

pub fn inst(&self, inst: Inst) -> &str

Stringifies an entire instruction. This includes the result if the instruction actually has one.

pub fn block(&self, bb: Block) -> &str

Stringifies a whole block. This includes the block label, any block parameters, and every instruction in the block.

pub fn func(&self, func: Func) -> &str

Stringifies a whole function. This includes every block, and the function prototype.

pub fn module(&self) -> &str

Returns the entire module as a string.