Struct sapphire::arena::ArenaMap

pub struct ArenaMap<K: ArenaKey, V> { /* private fields */ }

This is meant to act as a primary mapping of K -> V, where K is some key type and V is the value being stored. Other mappings that use the same key as an existing ArenaMap should use SecondaryMap instead.

This is effectively a typed wrapper around Vec<T>, the main advantage is that it does not implicitly convert into array types (i.e. it actually acts like a map instead of a sequence) and it only allows indexing with the correct type.

This allows type safety to be significantly increased, with unique key types for different types of collections. Key size can also be customized on a per-map basis, e.g. when a map is known to be small a smaller index type can be used.

Note that the key type puts an implicit limit on the size of the arena: a key type using u8 for example cannot store more than u8::MAX + 1 different objects in the arena, and trying to do so will panic.

arena_key! {
    struct Name;
}

let mut blocks = ArenaMap::new();
let bb: Name = blocks.insert("Hello!");

assert_eq!(blocks[bb], "Hello!");

Implementations

impl<K: ArenaKey, V> ArenaMap<K, V>

pub fn new() -> Self

Creates a new, empty arena. This creates the underlying Vec with Vec::default.

arena_key! { struct Key; }
let mut map = ArenaMap::new();
let k: Key = map.insert("Hello!");

pub fn with_capacity(capacity: usize) -> Self

Creates an empty arena with an initial capacity. This creates the underlying Vec with Vec::with_capacity.

let mut map = ArenaMap::with_capacity(10);
assert!(map.capacity() >= 10);

pub fn contains(&self, key: K) -> bool

Checks if the arena contains a given key, i.e. whether a given key has been returned from Self::insert at some point.

let mut map = ArenaMap::default();
let k1: Key = map.insert(true);
let k2 = map.next_key();
assert_eq!(map.contains(k1), true);
assert_eq!(map.contains(k2), false);

pub fn get(&self, key: K) -> Option<&V>

Accesses the arena and gets the value associated with a given key. If the key doesn’t exist (i.e. if Self::contains would have returned false, None is returned.

let mut map = ArenaMap::default();
let k1: Key = map.insert(15);
assert_eq!(map.get(k1), Some(&15));

pub fn get_mut(&mut self, key: K) -> Option<&mut V>

Accesses the arena and gets the value associated with a given key. If the key doesn’t exist (i.e. if Self::contains would have returned false, None is returned.

let mut map = ArenaMap::default();
let k1: Key = map.insert(15);
assert_eq!(map.get_mut(k1), Some(&mut 15));

pub fn insert(&mut self, value: V) -> K

Adds an item into the arena, and returns a key that can be used to access that data later.

let mut map = ArenaMap::default();
let k: Key = map.insert("Hello!");
assert_eq!(map[k], "Hello!");

pub fn next_key(&self) -> K

Gets the key that will be returned by Self::insert when it’s called next. This key is not valid until that Self::insert call occurs.

Once that Self::insert call occurs, the key is equivalent to using the key returned by Self::insert.

let mut map = ArenaMap::default();
let k1: Key = map.next_key();
assert_eq!(map.contains(k1), false);

let k2 = map.insert(0);
assert_eq!(map.contains(k1), true);
assert_eq!(map[k1], 0);
assert_eq!(k1, k2);

pub fn len(&self) -> usize

Gets the number of elements that have been pushed into the arena.

let mut map = ArenaMap::<Key, i32>::default();
assert_eq!(map.len(), 0);

pub fn is_empty(&self) -> bool

Checks if the arena has had any elements pushed into it.

let mut map = ArenaMap::<Key, i32>::default();
assert_eq!(map.is_empty(), true);

pub fn capacity(&self) -> usize

Returns the total number of elements that the arena can hold without needing to reallocate.

pub fn last(&self) -> Option<(K, &V)>

Returns the last element that was inserted in the map.

arena_key! { struct Key; }
let mut map = ArenaMap::default();
let k: Key = map.insert(15);
assert_eq!(map.last(), Some((k, &15)));

pub fn last_mut(&mut self) -> Option<(K, &mut V)>

Returns the last element that was inserted in the map.

arena_key! { struct Key; }
let mut map = ArenaMap::default();
let k: Key = map.insert(15);
assert_eq!(map.last_mut(), Some((k, &mut 15)));

pub fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more elements to be inserted. This may potentially reserve much more space depending on the state of the arena.

See Vec::reserve for the full behavior.

pub fn reserve_exact(&mut self, additional: usize)

Reserves the minimum capacity for exactly additional more elements to be inserted, without any potential over-reserving like Self::reserve does.

See Vec::reserve_exact for the full behavior.

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the arena as much as possible without removing any elements. The capacity may not actually be reduced to exactly len().

See Vec::shrink_to_fit for the full behavior.

pub fn keys(
&self
) -> impl Iterator<Item = K> + DoubleEndedIterator + ExactSizeIterator

Returns an iterator that iterates over the (valid) keys of the arena. Keys are guaranteed to be yielded in increasing order, in the range $$ [0, L) $$ such that $$ L $$ is len().

arena_key! { struct Key; }
let mut map = ArenaMap::default();
let k1: Key = map.insert(15);
let mut it = map.keys();
assert_eq!(it.next(), Some(k1));

pub fn values(&self) -> Iter<'_, V>

Returns an iterator that iterates over the values in the arena.

arena_key! { struct Key; }
let mut map = ArenaMap::default();
let k1: Key = map.insert(15);
let mut it = map.values();
assert_eq!(it.next(), Some(&15));

pub fn values_mut(&mut self) -> IterMut<'_, V>

Returns an iterator that iterates over the values in the arena, giving mutable references instead of shared references.

arena_key! { struct Key; }
let mut map = ArenaMap::default();
let k1: Key = map.insert(15);
let mut it = map.values_mut();
assert_eq!(it.next(), Some(&mut 15));

pub fn iter(
&self
) -> impl Iterator<Item = (K, &V)> + DoubleEndedIterator + ExactSizeIterator

Returns an iterator that iterates over the values in the arena, and the keys that map to those values.

arena_key! { struct Key; }
let mut map = ArenaMap::default();
let k1: Key = map.insert(15);
let mut it = map.iter();
assert_eq!(it.next(), Some((k1, &15)));

pub fn iter_mut(
&mut self
) -> impl Iterator<Item = (K, &mut V)> + DoubleEndedIterator + ExactSizeIterator

Returns an iterator that iterates over the values in the arena, and the keys that map to those values. Returns mutable references instead of shared references.

arena_key! { struct Key; }
let mut map = ArenaMap::default();
let k1: Key = map.insert(15);
let mut it = map.iter_mut();
assert_eq!(it.next(), Some((k1, &mut 15)));