Struct sapphire::arena::SecondaryMap

pub struct SecondaryMap<K: ArenaKey, V> { /* private fields */ }

Intended to be a dense secondary mapping K -> V for keys from a primary ArenaMap. This is to associate extra data with most (but ideally all) keys from a given PrimaryMap.

While the map does track insertion information in a dense bitset, this is not intended to map only a few keys, as this map stores space for every mapping just like ArenaMap does. If you anticipate only needing a few keys to be mapped and most not being mapped, use TODO instead.

The map also allows random insertion, but optimal performance is best obtained through using a combination of map_all_keys and map_some_keys.

This type of map does allow deletion of keys since the valid keys are not necessarily contiguous.

arena_key! { struct Player; }

// p1 => John, p2 => Bob
let mut players = ArenaMap::new();
let p1: Player = players.insert("John");
let p2 = players.insert("Bob");

// nice way to quickly do a map over all keys in `players`
let mut health = SecondaryMap::map_all_keys(&players, |_, _| 200);

// alternatively, you can do the insertions manually.
let mut ammo = SecondaryMap::with_primary(&players);

for k in players.keys() {
  ammo.insert(k, 50);
}

assert_eq!(players[p1], "John");
assert_eq!(health[p2], 200);
assert_eq!(ammo[p1], 50);

The underlying implementation is a Vec storing MaybeUninit<V>s along with information keeping track of which vec slots are inserted (and therefore initialized). That Vec expands if a key is inserted that doesn’t fit in the current Vec, this is what the “capacity” of the container is. The length is simply how many elements are initialized.

Implementations

impl<K: ArenaKey, V> SecondaryMap<K, V>

pub fn new() -> Self

Creates an empty map with 0 as the capacity.

arena_key! { struct Key; }
let map = SecondaryMap::<Key, i32>::new();

pub fn with_capacity(cap: usize) -> Self

Creates an empty map that is pre-allocated for a specific number of keys.

See Self::reserve for details on how capacity works in a SecondaryMap.

arena_key! { struct Key; }
let primary = ArenaMap::<Key, usize>::new();
let map = SecondaryMap::<Key, i32>::with_capacity(primary.len());

pub fn with_primary<T>(primary: &ArenaMap<K, T>) -> Self

Creates an empty map that is configured for optimal performance when used with keys from a specific primary ArenaMap.

See Self::reserve for details on how capacity works in a SecondaryMap.

arena_key! { struct Key; }
let primary = ArenaMap::<Key, usize>::new();
let map = SecondaryMap::<Key, i32>::with_primary(&primary);

pub fn map_all_keys<T, F>(primary: &ArenaMap<K, T>, f: F) -> Selfwhere
F: FnMut(&ArenaMap<K, T>, K) -> V,

Efficiently creates a SecondaryMap by applying a function to each key inside of primary and inserting the result.

Effectively, this results on a mapping of K -> f(K).

arena_key! { struct Key; }
let mut primary = ArenaMap::new();
let k1: Key = primary.insert(4);
let k2 = primary.insert(16);

let map = SecondaryMap::map_all_keys(&primary, |map, key| map[key] * 2);

assert_eq!(map[k1], 8);
assert_eq!(map[k2], 32);

pub fn map_some_keys<T, F>(primary: &ArenaMap<K, T>, f: F) -> Selfwhere
F: FnMut(&ArenaMap<K, T>, K) -> Option<V>,

Efficiently creates a SecondaryMap by applying a function to each key inside of primary and inserting the result. The difference between this and Self::map_all_keys is that the mapping can skip over some keys.

Effectively, this results on a mapping of K -> f(K) for all K where f(K) != None

arena_key! { struct Key; }
let mut primary = ArenaMap::new();
let k1: Key = primary.insert(4);
let k2 = primary.insert(5);    

let map = SecondaryMap::map_some_keys(&primary, |map, key| {
    match map[key] {
        x if x % 2 == 0 => None,
        x => Some(x * 2)
    }   
});

assert_eq!(map[k2], 10);
assert!(!map.contains(k1));

pub fn contains(&self, key: K) -> bool

Checks if the map contains a value for a given key.

arena_key! { struct Key; }
let mut primary = ArenaMap::new();
let mut secondary = SecondaryMap::new();
let k1: Key = primary.insert("Hello!");
secondary.insert(k1, 16);

assert!(secondary.contains(k1));

pub fn insert(&mut self, key: K, value: V) -> Option<V>

Inserts a key/value pair into the table, mapping key -> value.

If there is already a value mapped to key in the table, that value is returned. Otherwise, None is returned.

arena_key! { struct Key; }
let mut primary = ArenaMap::new();
let mut secondary = SecondaryMap::new();
let k1: Key = primary.insert("Hello!");

assert_eq!(secondary.insert(k1, 16), None); // no previous value
assert_eq!(secondary.insert(k1, 13), Some(16)); // got previous

pub fn get(&self, key: K) -> Option<&V>

Gets a value out of the map if the key has been previously added with Self::insert.

arena_key! { struct Key; }
let mut primary = ArenaMap::new();
let mut secondary = SecondaryMap::new();
let k1: Key = primary.insert("Hello!");

secondary.insert(k1, 13);

assert_eq!(secondary.get(k1), Some(&13));

pub fn get_mut(&mut self, key: K) -> Option<&mut V>

Gets a value out of the map if the key has been previously added with Self::insert.

arena_key! { struct Key; }
let mut primary = ArenaMap::new();
let mut secondary = SecondaryMap::new();
let k1: Key = primary.insert("Hello!");

secondary.insert(k1, 13);

assert_eq!(secondary.get_mut(k1), Some(&mut 13));

pub fn take(&mut self, key: K) -> Option<V>

Removes a key from the table and returns the value that was at that key, if it existed. If the key isn’t in the table, None is returned.

arena_key! { struct Key; }
let mut primary = ArenaMap::new();
let mut secondary = SecondaryMap::new();
let k1: Key = primary.insert("Hello!");
secondary.insert(k1, 13);

assert_eq!(secondary.take(k1), Some(13));
assert_eq!(secondary.take(k1), None);

pub fn remove(&mut self, key: K) -> bool

Removes a key from the table if it exists. Returns whether or not anything was actually removed.

arena_key! { struct Key; }
let mut primary = ArenaMap::new();
let mut secondary = SecondaryMap::new();
let k1: Key = primary.insert("Hello!");
secondary.insert(k1, 13);

assert_eq!(secondary.remove(k1), true);
assert_eq!(secondary.contains(k1), false);

pub fn reserve(&mut self, additional: usize)

Informs the map of how big the primary arena is.

Unlike a Vec, the capacity is not quite the number of elements that can be stored without re-allocating, although it often is. The capacity is a measure of the highest key value that can be inserted without reallocating, i.e. how big the map thinks that the primary arena is.

If the primary map has gained more elements, the secondary map should have that new length reserved.

arena_key! { struct Key; }
let primary = ArenaMap::<Key, i32>::new();
let mut map = SecondaryMap::<Key, isize>::new();
map.reserve(primary.len());

pub fn shrink_to_fit(&mut self)

Shrinks the underlying capacity of the arena as much as possible without removing any slots. Note that this will not reduce Self::capacity, as a SecondaryMap’s capacity is not the same thing as a Vec’s capacity. See Self::reserve for more details.

See Vec::shrink_to_fit for the full behavior.

pub fn capacity(&self) -> usize

Returns how big the map believes that the primary arena is.

See Self::reserve for details on what exactly this means.

pub fn len(&self) -> usize

Returns the number of values that have been inserted into this map, i.e. the number of keys that have associated values.

arena_key! { struct Key; }
let mut map = SecondaryMap::<Key, i32>::new();
assert_eq!(map.len(), 0);

pub fn is_empty(&self) -> bool

Checks if the map has any values in it, i.e. if any key/value pairs have been inserted.

arena_key! { struct Key; }
let mut map = SecondaryMap::<Key, i32>::new();
assert!(map.is_empty());

pub fn keys(&self) -> impl Iterator<Item = K> + '_

Returns an iterator that iterates over the (valid) keys of the arena. Keys are guaranteed to be yielded in increasing order, in the range $$ [0, L) $$ such that $$ L $$ is len().

arena_key! { struct Key; }
let mut map = ArenaMap::default();
let k1: Key = map.insert(15);
let mut it = map.keys();
assert_eq!(it.next(), Some(k1));

pub fn values(&self) -> impl Iterator<Item = &V>

Returns an iterator that iterates over the values in the arena.

arena_key! { struct Key; }
let mut map = ArenaMap::default();
let k1: Key = map.insert(15);
let secondary = SecondaryMap::map_all_keys(&map, |m, k| m[k] * 2);
let mut it = secondary.values();
assert_eq!(it.next(), Some(&30));

pub fn values_mut(&mut self) -> impl Iterator<Item = &mut V>

Returns an iterator that iterates over the values in the arena, giving mutable references instead of shared references.

arena_key! { struct Key; }
let mut map = ArenaMap::default();
let k1: Key = map.insert(15);
let mut secondary = SecondaryMap::map_all_keys(&map, |m, k| m[k] * 2);
let mut it = secondary.values_mut();
assert_eq!(it.next(), Some(&mut 30));

pub fn iter(&self) -> impl Iterator<Item = (K, &V)>

Returns an iterator that iterates over the values in the arena, and the keys that map to those values.

arena_key! { struct Key; }
let mut map = ArenaMap::default();
let k1: Key = map.insert(15);
let mut secondary = SecondaryMap::with_primary(&map);
secondary.insert(k1, 'c');
let mut it = secondary.iter();
assert_eq!(it.next(), Some((k1, &'c')));

pub fn iter_mut(&mut self) -> impl Iterator<Item = (K, &mut V)>

Returns an iterator that iterates over the values in the arena, and the keys that map to those values. Returns mutable references instead of shared references.

arena_key! { struct Key; }
let mut map = ArenaMap::default();
let k1: Key = map.insert(15);
let mut it = map.iter_mut();
assert_eq!(it.next(), Some((k1, &mut 15)));