Struct sapphire::arena::SecondarySet

pub struct SecondarySet<K: ArenaKey> { /* private fields */ }

Intended to be a dense representation for a set of keys from a primary map.

This is theoretically equivalent to a SecondaryMap<K, ()> but is a bit leaner under-the-hood, being implemented as just a bitvector with no additional slot storage.

arena_key! { struct Key; }
let mut map = ArenaMap::default();
let k1: Key = map.insert(15);
let k2 = map.insert(20);
let secondary = SecondarySet::map_keys(&map, |m, k| m[k] % 2 == 0);

assert_eq!(secondary.contains(k1), false);
assert_eq!(secondary.contains(k2), true);

Implementations

impl<K: ArenaKey> SecondarySet<K>

pub fn new() -> Self

Creates an empty set with 0 as the capacity.

arena_key! { struct Key; }
let map = SecondarySet::<Key>::new();

pub fn with_capacity(cap: usize) -> Self

Creates an empty set that is pre-allocated for a specific number of keys.

arena_key! { struct Key; }
let primary = ArenaMap::<Key, usize>::new();
let map = SecondarySet::<Key>::with_capacity(primary.len());

pub fn with_primary<T>(primary: &ArenaMap<K, T>) -> Self

Creates an empty map that is configured for optimal performance when used with keys from a specific primary ArenaMap.

arena_key! { struct Key; }
let primary = ArenaMap::<Key, usize>::new();
let map = SecondarySet::<Key>::with_primary(&primary);

pub fn map_keys<T, F>(primary: &ArenaMap<K, T>, f: F) -> Selfwhere F: FnMut(&ArenaMap<K, T>, K) -> bool,

Efficiently creates a SecondarySet by applying a function to each key inside of primary to determine if a key should be included in the set or not.

Effectively, this results in the set $$ { k | f(k) is true } $$

arena_key! { struct Key; }
let mut primary = ArenaMap::<Key, i32>::new();
let k1 = primary.insert(4);
let k2 = primary.insert(5);

let is_even = SecondarySet::map_keys(&primary, |map, key| map[key] % 2 == 0);

assert_eq!(is_even.contains(k1), true);
assert_eq!(is_even.contains(k2), false);

pub fn cardinality(&self) -> usize

Returns the number of keys that are currently in the set.

arena_key! { struct Key; }
let mut map = ArenaMap::default();
let k1: Key = map.insert(15);
let k2 = map.insert(20);
let secondary = SecondarySet::map_keys(&map, |m, k| {
eprintln!("{}", m[k]);
m[k] % 2 == 0});
assert_eq!(secondary.cardinality(), 1);

pub fn is_empty(&self) -> bool

Returns if the set is completely empty.

arena_key! { struct Key; }
let mut map = ArenaMap::default();
let k1: Key = map.insert(15);
let k2 = map.insert(17);
let secondary = SecondarySet::map_keys(&map, |m, k| m[k] % 2 == 0);
assert_eq!(secondary.is_empty(), true);

pub fn capacity(&self) -> usize

Returns how many keys can be stored in the set without reallocating.

pub fn contains(&self, key: K) -> bool

Returns whether or not a key is present in the set.

arena_key! { struct Key; }
let mut map = ArenaMap::default();
let k1: Key = map.insert(15);
let k2 = map.insert(20);
let mut secondary = SecondarySet::default();
assert_eq!(secondary.insert(k1), false);
assert_eq!(secondary.contains(k1), true);

pub fn insert(&mut self, key: K) -> bool

Inserts a key into the set, returns whether the key was in the set prior to insertion.

arena_key! { struct Key; }
let mut map = ArenaMap::default();
let k1: Key = map.insert(15);
let k2 = map.insert(20);
let mut secondary = SecondarySet::default();
assert_eq!(secondary.insert(k1), false);
assert_eq!(secondary.contains(k1), true);

pub fn remove(&mut self, key: K) -> bool

Removes a key from the set if it’s present, and returns whether or not a key was removed.

arena_key! { struct Key; }
let mut map = ArenaMap::default();
let k1: Key = map.insert(15);
let k2 = map.insert(20);
let mut secondary = SecondarySet::default();
secondary.insert(k1);
assert_eq!(secondary.remove(k1), true);
assert_eq!(secondary.remove(k1), false);

pub fn clear(&mut self)

Removes all keys from the set without deallocating the buffer.

arena_key! { struct Key; }
let mut map = ArenaMap::default();
let k1: Key = map.insert(15);
let k2 = map.insert(20);
let mut secondary = SecondarySet::default();
secondary.insert(k1);
secondary.clear();
assert_eq!(secondary.contains(k1), false);

pub fn keys(&self) -> impl Iterator<Item = K> + '_

Returns an iterator that iterates over the keys that are present in the set. Keys are guaranteed to be yielded in increasing order, in the range $$ [0, C) $$ such that $$ C $$ is cardinality().

arena_key! { struct Key; }
let mut map = ArenaMap::default();
let k1: Key = map.insert(15);
let k2 = map.insert(20);
let secondary = SecondarySet::map_keys(&map, |m, k| m[k] % 2 != 0);
let mut it = secondary.keys();
assert_eq!(it.next(), Some(k1));
assert_eq!(it.next(), None);