Struct sapphire::arena::UniqueArenaMap

pub struct UniqueArenaMap<K, V>where
    K: ArenaKey,
    V: Eq + Hash,{ /* private fields */ }

Contains a table of immutable, unique elements. All elements are only stored once, no duplicates are stored inside of the arena itself. Lookups by key are almost as efficient as they are in a plain ArenaMap, although an extra pointer dereference must be done.

This deduplication is paid for by adding hashing overhead to every call to Self::insert, and by requiring values to be stored in an Rc<V>. This makes them immutable, as allowing them to be mutable would invalidate hashes.

A very similar API to ArenaMap is exposed.

Implementations

impl<K, V> UniqueArenaMap<K, V>where
K: ArenaKey,
V: Hash + Eq,

pub fn new() -> Self

Creates a new arena with nothing in it.

pub fn with_capacity(capacity: usize) -> Self

Creates the arena with a starting capacity.

See Vec::with_capacity for details.

pub fn insert(&mut self, value: V) -> K

Inserts a value that is either unique or a duplicate of some value that is already in the arena. If value is a duplicate, value is dropped and a key that refers to an equivalent value (i.e. value that compares equal via Eq and has the same hash via Hash) is returned.

arena_key! { struct Key; }
let rc = Rc::new(42);
let mut map = UniqueArenaMap::new();
let k1: Key = map.insert(rc.clone()); // original, inserted
let k2: Key = map.insert(rc.clone()); // duplicate, not inserted
assert_eq!(k1, k2);
assert_eq!(Rc::strong_count(&rc), 2); // only 2 refs, `rc` and 1x copy inside the map

pub fn insert_clone_if(&mut self, value: &V) -> Kwhere
V: Clone,

Equivalent to Self::insert but will only make a copy of the value if the value does not exist in the map. This requires that V implements Clone.

pub fn contains(&self, key: K) -> bool

Checks if the arena contains a given key, i.e. whether a given key has been returned from Self::insert at some point.

let mut map = UniqueArenaMap::default();
let k1: Key = map.insert(true);
assert_eq!(map.contains(k1), true);

pub fn get(&self, key: K) -> Option<&V>

Accesses the arena and gets the value associated with a given key. If the key doesn’t exist (i.e. if Self::contains would have returned false, None is returned.

let mut map = UniqueArenaMap::default();
let k1: Key = map.insert(15);
assert_eq!(map.get(k1), Some(&15));

pub fn next_key(&self) -> K

Gets the key that will be returned by Self::insert when it’s called next (with a non-duplicate key). This key is not valid until that Self::insert call occurs.

Once that Self::insert call occurs, the key is equivalent to using the key returned by Self::insert.

let mut map = UniqueArenaMap::default();
let k1: Key = map.next_key();
assert_eq!(map.contains(k1), false);

let k2 = map.insert(0);
assert_eq!(map.contains(k1), true);
assert_eq!(map[k1], 0);
assert_eq!(k1, k2);

pub fn len(&self) -> usize

Gets the number of elements that have been pushed into the arena.

let mut map = UniqueArenaMap::<Key, i32>::default();
assert_eq!(map.len(), 0);

pub fn is_empty(&self) -> bool

Checks if the arena has had any elements pushed into it.

let mut map = UniqueArenaMap::<Key, i32>::default();
assert_eq!(map.is_empty(), true);

pub fn capacity(&self) -> usize

Returns the total number of elements that the arena can hold without needing to reallocate.

pub fn last(&self) -> Option<(K, &V)>

Returns the last element that was inserted in the map. Note that this is not necessarily the last call to Self::insert, this is the last element that was actually inserted (i.e. the last non-duplicate element to be inserted).

arena_key! { struct Key; }
let mut map = UniqueArenaMap::default();
let k: Key = map.insert(15);
assert_eq!(map.last(), Some((k, &15)));

pub fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more unique elements to be inserted. This may potentially reserve much more space depending on the state of the arena.

See Vec::reserve for the full behavior.

pub fn reserve_exact(&mut self, additional: usize)

Reserves the minimum capacity for exactly additional more elements to be inserted, without any potential over-reserving like Self::reserve does.

See Vec::reserve_exact for the full behavior.

pub fn shrink_to_fit(&mut self)

Shrinks the capacity of the arena as much as possible without removing any elements. The capacity may not actually be reduced to exactly len().

See Vec::shrink_to_fit for the full behavior.

pub fn keys(&self) -> impl Iterator<Item = K>

Returns an iterator that iterates over the (valid) keys of the arena. Keys are guaranteed to be yielded in increasing order, in the range $$ [0, L) $$ such that $$ L $$ is len().

arena_key! { struct Key; }
let mut map = UniqueArenaMap::default();
let k1: Key = map.insert(15);
let mut it = map.keys();
assert_eq!(it.next(), Some(k1));

pub fn values(&self) -> impl Iterator<Item = &V>

Returns an iterator that iterates over the values in the arena.

arena_key! { struct Key; }
let mut map = UniqueArenaMap::default();
let k1: Key = map.insert(15);
let mut it = map.values();
assert_eq!(it.next(), Some(&15));

pub fn iter(&self) -> impl Iterator<Item = (K, &V)>

Returns an iterator that iterates over the values in the arena, and the keys that map to those values.

arena_key! { struct Key; }
let mut map = UniqueArenaMap::default();
let k1: Key = map.insert(15);
let mut it = map.iter();
assert_eq!(it.next(), Some((k1, &15)));