Module sapphire::cli

Contains utility code specifically for the CLI tools located in the tools/ subdirectory.

All of these tools have similar command-line arguments and they all should look/feel uniform, so most of the code is pulled into this module and then used in the drivers of the different tools.

Structs

BaseOptions

Basic options that every CLI tool in the suite takes in.

Enums

IRFormat

The format for a tool emitting SIR to emit in.

MachineFormat

The format for a tool emitting native code to emit in.

Functions

default

Gets the baseline default options that every tool needs.

emit_machine_format

An emit argument for tools that emit machine-specific code.

emit_sir

Gets the emit format for a tool that emits SIR

inputs

Gets the input file specified on the CLI.

jobs

Gets the number of concurrent threads to use for a given task

output

Gets the output file specified on the CLI, if one exists.

tool

Returns a [bpaf::OptionParser] preconfigured with the standard Sapphire options and nothing else.

tool_with

Returns a [bpaf::OptionParser] preconfigured with the standard Sapphire options and additional tool-specific options.

verbose

Checks for the presence of -v or --verbose