Module sapphire::codegen

APIs for the compiler back-end and code-generation facilities

These APIs mostly abstract away any target-specific details, but there is target-specific code in this module. Anything in a module with an architecture in its name (e.g. sapphire::codegen::x86_64::* or sapphire::codegen::aarch64::*) are CPU-specific.

Modules

patterns

This module defines SIR matchers that can be used inside of instruction selection code.

x86_64

x86-64 Backend

Structs

AggregateLayout

Models the more complex layout information

Allocation

A representation of a valid allocation for a MIR program.

AvailableRegisters

A representation of the specific available registers on a particular ABI.

Backend

A generic backend using a specified emitter. The Backend doesn’t directly perform instruction selection by itself, that is done during the creation of the Backend instance.

CallUseDefId

Used by StackFrame::register_use_def_call and StackFrame::call_use_defs as a unique identifier for a call in a function.

CodegenOptions

A set of detailed codegen configuration options that are provided during the lowering process.

ConservativeLiveIntervals

Models a simpler form of live-ness, live intervals.

FixedIntervals

A list of “fixed” intervals.

GenericISel

The main instruction selection pipeline that takes in a target-specific instruction selector and repeatedly invokes it to lower each instruction.

GraphColoringRegAlloc

A register allocator using the “iterated register coalescing” algorithm

LegalizeAggregatesForABI

Demotion pass that moves any aggregates (arrays, structures) that are “too big” for a specific arch/abi configuration to deal with in registers onto the stack.

LinearScanRegAlloc

A linear scanning based register allocator

LiveInterval

A live interval denoting the instruction a register is first defined by and the instruction a register is last used by.

LiveRanges

Models the live range information necessary for the register allocators.

LoweringContext

The “context” of a specific lowering.

MIRBlock

A reference to a single block of MIR.

MIRBlockInterval

Defines a “block” in MIR. This is just a range that maps to a sublist in the overall list of instructions in a function.

MIRFunction

A single function in MIR.

MIRModule

Equivalent of a Module for MIR.

PReg

Represents a single physical register on a CPU. The register class is stored in one bit, while the register number is in the other bits that are below that bit.

PresetBackends

A set of pre-made backend pipelines for various architectures and ABIs, made for convenience.

PresetTargets

Preset target configurations for specific architecture/ABI sets.

ProgramPoint

A single point in a MIR program.

ProgramPoints

A nice way to iterate over program points when desired.

Reg

A register that is either a virtual register or a physical register of a given register class.

RegToRegCopy

An abstract representation of a mov between two registers

RegisterMapping

Models mapping an instruction to the register allocation it has.

Rewriter

Performs the final phase of codegen, rewriting register-allocated x86-64 MIR to be fully valid x86-64 instructions that are ready to emit.

StackRegAlloc

A “register allocator” that puts everything on the stack, only loading values to perform operations.

Target

Models the data-layout, calling conventions and other details necessary for generating code for a specific architecture and ABI.

TypeLayout

Models the layout of a given type. This is the size and alignment requirements of the type, if the size/alignment/offset of subfields is desired you either need to get the layout of that subfield’s type, or use the AggregateLayout::offset API.

UnconditionalBranch

An abstract representation of a jmp that will always be taken

VReg

Represents a single virtual register. The register class is stored in one bit, while the register number is in the other bits that are below that bit.

WriteableReg

A new-type for a register that is writeable. This doesn’t actually enforce anything but makes assumptions obvious in the code since the conversion is explicit.

Enums

CPUArch

Models the specific CPU architecture that is being targeted.

Extern

The different types of external entities that MIR can reference. This is here for emitting purposes, so the symbols that the linker needs to resolve can be listed.

ParallelCopyLocation

A location used in parallel copy scheduling.

RegClass

Models the type of register that a given register is.

SpillReload

Models a single spill or reload generated by the register allocator.

TargetPair

Supported CPU/OS targets. These are associated with their default ABIs on the platforms being referred to.

VariableLocation

The location of a single “variable.” This denotes something at the ABI level, e.g. stackslots, parameters and the like. This identifies where they are in a way that the code generator can understand.

Traits

Architecture

Models the architecture-specific details that a backend needs to deal with, mainly fundamental type layouts.

CallingConv

Models a calling convention.

Emitter

An emitter that can emit different formats from a given type of MachInst.

InstructionSelector

An instruction selector that lowers one Inst into (potentially many) MachInsts.

MachInst

Defines the API for a machine instruction.

Platform

Models the idea of a “platform,” i.e. a target operating environment that code is being generated for.

ProgramPointsIterator

A nice way to iterate over program points when desired.

RegisterAllocator

Models a register allocator that works over MIR.

StackFrame

Interface for a generic “stack frame” that a target can implement. This is used by the code generator to place data in the correct place at the function level.

Functions

schedule_parallel_copies

Given a list of block parameters and arguments being passed as the new values of those block parameters, this schedules them correctly in a way that doesn’t fall into the swap problem.

Type Definitions

Ctx

The context passed to any code doing instruction selection

FramelessCtx

Context passed to instruction selection code that’s actually a part of StackFrame.

FuncFramePair

Holds a MIRFunction along with the associated stack frame information for it.

RegCollector

A collector that can have registers added to it.