Module sapphire::codegen::patterns

This module defines SIR matchers that can be used inside of instruction selection code.

These matchers are not meant to be used in optimization passes, see the matchers module for those. These are meant to work with codegen infrastructure.

Structs

BasicISelMergeMatcher

A matcher that wraps up a matcher function.

Traits

ISelMergeMatcher

A basic matcher for a single instruction, for use in merging during instruction selection.

Functions

and

Matches against and instructions, ignoring operands

and_with

Matching against and instructions and their operands

any

Matches any instruction, or anything else that produces a value.

ashr

Matches against ashr instructions, ignoring operands

ashr_with

Matching against ashr instructions and their operands

fadd

Matches against fadd instructions, ignoring operands

fadd_with

Matching against fadd instructions and their operands

fdiv

Matches against fdiv instructions, ignoring operands

fdiv_with

Matching against fdiv instructions and their operands

fmul

Matches against fmul instructions, ignoring operands

fmul_with

Matching against fmul instructions and their operands

frem

Matches against frem instructions, ignoring operands

frem_with

Matching against frem instructions and their operands

fsub

Matches against fsub instructions, ignoring operands

fsub_with

Matching against fsub instructions and their operands

iadd

Matches against iadd instructions, ignoring operands

iadd_with

Matching against iadd instructions and their operands

icmp

Matches on any icmp instruction without matching on any of its operands.

icmp_with

Matches on an icmp instruction with a specific condition, and two matching operands.

iconst

Matches any iconst instruction.

iconst_of

Matches on an iconst instruction, and sets val to the value the iconst produces.

iconst_val

Matches any iconst instruction with any value.

imm32

Matches any iconst instruction that has a value that can fit in an imm32

imul

Matches against imul instructions, ignoring operands

imul_with

Matching against imul instructions and their operands

isub

Matches against isub instructions, ignoring operands

isub_with

Matching against isub instructions and their operands

load

Matches any load instruction.

lshr

Matches against lshr instructions, ignoring operands

lshr_with

Matching against lshr instructions and their operands

mark_merged

Used when you want to actually perform the merge between base and its operands after successfully matching via matches().

mark_merged_operand

Used when you want to actually perform the merge onto base with the operand val after successfully matching via matches_operand.

matches

Used when you want to check if a merge between base and its operands is safe, this must be followed by a mark_merged to actually complete the merge.

matches_operand

Used when you want to check if a merge between base and its operands is safe, this must be followed by a mark_merged_operand to actually complete the merge.

merge_if_matches

Performs the two-step “check if able to merge” and “mark the merge as done” process in one step.

merge_if_matches_operand

Performs the two-step “check if able to merge” and “mark the merge as done” process in one step.

neg1

Matches an integral constant of -1 for any integer type

or

Matches against or instructions, ignoring operands

or_with

Matching against or instructions and their operands

phi

Matches any block parameter (aka phi nodes).

sdiv

Matches against sdiv instructions, ignoring operands

sdiv_with

Matching against sdiv instructions and their operands

shl

Matches against shl instructions, ignoring operands

shl_with

Matching against shl instructions and their operands

srem

Matches against srem instructions, ignoring operands

srem_with

Matching against srem instructions and their operands

udiv

Matches against udiv instructions, ignoring operands

udiv_with

Matching against udiv instructions and their operands

urem

Matches against urem instructions, ignoring operands

urem_with

Matching against urem instructions and their operands

xor

Matches against xor instructions, ignoring operands

xor_with

Matching against xor instructions and their operands

zero

Matches any instruction that yields a constant 0 value