Function sapphire::codegen::patterns::matches

pub fn matches<Arch, Matcher>(
    base: Inst,
    ctx: FramelessCtx<'_, '_, '_, Arch>,
    matcher: &Matcher
) -> boolwhere
    Arch: Architecture,
    Matcher: ISelMergeMatcher<Arch>,

Used when you want to check if a merge between base and its operands is safe, this must be followed by a mark_merged to actually complete the merge.

Usually you don’t want this, use the single step merge_if_matches_operand instead unless you specifically need to run additional checks outside the pattern syntax.