Trait sapphire::codegen::patterns::ISelMergeMatcher

pub trait ISelMergeMatcher<Arch>where
    Arch: Architecture,{
    // Required methods
    fn matches_as_base(
        &self,
        base: Inst,
        ctx: FramelessCtx<'_, '_, '_, Arch>
    ) -> bool;
    fn matches_as_operand(
        &self,
        val: Value,
        base: Inst,
        ctx: FramelessCtx<'_, '_, '_, Arch>
    ) -> bool;
    fn mark_as_merged_base(
        &self,
        base: Inst,
        ctx: FramelessCtx<'_, '_, '_, Arch>
    );
    fn mark_as_merged_operand(
        &self,
        val: Value,
        base: Inst,
        ctx: FramelessCtx<'_, '_, '_, Arch>
    );
}

A basic matcher for a single instruction, for use in merging during instruction selection.

This uses the checks inside the LoweringContext to see not only if the IR matches the pattern, but to see if the match is able to be merged.

Required Methods

fn matches_as_base( &self, base: Inst, ctx: FramelessCtx<'_, '_, '_, Arch> ) -> bool

Runs the matcher against an instruction, the instruction as treated as the base of the merge (and is thus not subject to the requirements in LoweringContext::able_to_merge_with).

All of the operands of base (that are matched on) are subject to the requirements.

fn matches_as_operand( &self, val: Value, base: Inst, ctx: FramelessCtx<'_, '_, '_, Arch> ) -> bool

Runs the matcher against val, the instruction as treated as an operand of the merge and is checked against base using LoweringContext::able_to_merge_with.

All of the operands of val (that are matched on) are also subject to the requirements.

fn mark_as_merged_base(&self, base: Inst, ctx: FramelessCtx<'_, '_, '_, Arch>)

Once Self::matches_as_base has been called and returns true, this can be used to actually go update the state of the LoweringContext and mark that each matched operand is going to be merged with base

fn mark_as_merged_operand( &self, val: Value, base: Inst, ctx: FramelessCtx<'_, '_, '_, Arch> )

Once Self::matches_as_operand has been called and returns true, this can be used to actually go update the state of the LoweringContext and mark that each matched instruction is going to be merged with base