Struct sapphire::codegen::LoweringContext

pub struct LoweringContext<'module, 'target, Arch: Architecture> {
    pub target: &'target Target<Arch>,
    pub module: &'module Module,
    pub options: CodegenOptions,
    /* private fields */
}

The “context” of a specific lowering.

This is given to a specific selector when its asked to lower an instruction, this is all the information it needs to update when doing so.

Fields
target: &'target Target<Arch>

The target that code is being generated for. This should be queried by the selector.

module: &'module Module

The module being lowered. This is here to refer to module-scoped information

options: CodegenOptions

Codegen options, available for instruction selection code to access

Implementations

impl<'module, 'target, Arch: Architecture> LoweringContext<'module, 'target, Arch>

pub fn new_for( target: &'target mut Target<Arch>, module: &'module Module ) -> Self

Creates a LoweringContext prepared for a specific target and module.

pub fn reference_external_func(&mut self, func: Func) -> Str

Marks that the module references func as an external, and then returns a Str that refers to that function’s name.

pub fn prepare_for_func( &mut self, func: &'module Function ) -> (Str, ArenaMap<MIRBlock, MIRBlockInterval>)

Prepares the LoweringContext to start having instructions from func be emitted.

Returns the function name and block lookup, both suitable for giving to a MIRFunction.

This prepares the internal state for func, and computes things like usage data, instruction colors, and the block interval lookup.

pub fn take_insts(&mut self) -> Vec<Arch::Inst>

Takes the current instruction buffer, emptying it in the process. The vec returned will be the contents of the function emitted so far, in program order.

pub fn emit(&mut self, inst: Arch::Inst)

Push an instruction into the result buffer.

This should be called in the order that the instructions should appear in program order, meaning that if you want add to be followed by a mul in the final program, you should call this with add and then call it again with mul after.

pub fn begin_linear_lowering(&mut self)

Called before we begin lowering a single ir::Inst. This must be accompanied by a call to Self::end_linear_lowering after the lowering.

This prepares the internal buffer so that the lowering code can emit MachInsts in program order while the ir::Insts are being lowered in reverse order.

pub fn end_linear_lowering(&mut self)

Finalizes the process began by Self::begin_linear_lowering. Inserts the instructions emitted by a single lowering into the MIR buffer.

pub fn begin_fixed_interval(&mut self, reg: PReg)

Begins creating a fixed interval for each register in regs.

This must be accompanied by a call to Self::end_fixed_interval once the instructions included in the interval are emitted.

pub fn begin_fixed_intervals(&mut self, regs: &[PReg])

Begins creating a fixed interval for each register in regs.

This must be accompanied by a call to Self::end_fixed_interval once the instructions included in the interval are emitted.

pub fn begin_caller_defined_fixed_interval(&mut self, reg: PReg)

This is specifically for lowering parameters and similar caller-defined registers, where the interval needs to start before the function.

Other than that, it works the same way as Self::begin_fixed_interval.

pub fn end_fixed_interval(&mut self, reg: PReg, insts_after_interval: usize)

Finishes creating a fixed interval for reg.

It is assumed that all the instructions emitted via Self::emit since the call to Self::begin_fixed_interval are included in the interval.

If there will be any instructions emitted after this interval, the exact number needs to be given in insts_after_interval.

This must be accompanied by a call to Self::begin_fixed_interval before this is called, and that call must have the same registers.

pub fn end_fixed_intervals( &mut self, regs: &[PReg], insts_after_interval: usize )

Finishes creating a fixed interval for each register in regs.

It is assumed that all the instructions emitted via Self::emit since the call to Self::begin_fixed_interval are included in the interval.

If there will be any instructions emitted after this interval, the exact number needs to be given in insts_after_interval.

This must be accompanied by a call to Self::begin_fixed_interval before this is called, and that call must have the same registers.

pub fn end_caller_defined_fixed_interval( &mut self, reg: PReg, insts_after_interval: usize )

This is specifically for lowering parameters and similar caller-defined registers, where the interval needs to start before the function.

Other than that, it works the same way as Self::end_fixed_interval.

pub fn resolve(&mut self, string: &str) -> Str

Resolves (and possibly inserts) a string in the MIR module’s string pool.

pub fn mir_block(&self, block: Block) -> MIRBlock

Finds the MIRBlock associated with a given Block.

pub fn able_to_merge_with(&self, merge: Inst, base: Inst) -> bool

Checks if the instruction merge can be merged with base during selection.

Two instructions can only be merged if one of the following conditions are true AND the instruction only has one use in total:

The instruction to be merged is a constant of some sort

The instructions have the same side-effect ‘color’

pub fn mark_merged_with(&mut self, merge: Inst, base: Inst)

Marks that the instruction merge was merged with base. This is only to be called when it has already been determined that the merge is safe.

Once this is called, merge will not be lowered later.

pub fn next_vreg(&mut self, class: RegClass) -> Reg

Gets a new vreg for a given register class. This register is unique, it will not be returned again (and therefore must be kept if you want to use it multiple times).

pub fn constrain_result(&mut self, value: Value, reg: Reg)

Enforce a constraint that values result be placed in reg. When the instruction yielding that value is emitted, the selector must follow this constraint.

If a value does not have a constraint placed on it, it can be placed in any result (although any instruction without a side effect can also be skipped if its result is not constrained).

The intent is that whenever a value is used, the instruction using it selects a register and constrains the value to be put in that register, and then once that instruction is emitted later it will be

pub fn result_reg(&mut self, value: Value, class: RegClass) -> Reg

Gets a result register for a given value.

If the value has a constraint placed on it, that register is returned. Otherwise, a new register is selected, the value is constrained to that register, and that register is returned.

pub fn maybe_result_reg(&mut self, value: Value) -> Option<Reg>

If a value has a constrained result register, returns that. If it does not, this returns None.

This is only here for instructions like icmp and fcmp that are not fully merged with branches, other instructions should not use this.

pub fn func(&self) -> &'module Function

Gets the current function being lowered.

pub fn is_merged(&self, inst: Inst) -> bool

Checks if a given instruction has been merged with another or not

pub fn val_class(val: Value, def: &FunctionDefinition) -> RegClass

Gets the RegClass that a given Value can go in.