Struct sapphire::codegen::Target

pub struct Target<Arch: Architecture> { /* private fields */ }

Models the data-layout, calling conventions and other details necessary for generating code for a specific architecture and ABI.

Implementations

impl<Arch: Architecture> Target<Arch>

pub fn prepare_for(&mut self, module: &Module)

Calculates all the type layouts for every type present in a given module.

This must be called before the target can be queried for the layouts of arbitrary types.

pub fn layout_of(&self, ty: Type) -> TypeLayout

Gets the TypeLayout associated with a given type.

pub fn new_frame(&self, func: &Function) -> Box<dyn StackFrame<Arch>>

Gets a new stack frame for a given function

pub fn new_callcc(&self) -> &'static dyn CallingConv<Arch>

Gets a new calling convention for a given function

pub fn cpu(&self) -> CPUArch

Returns the specific CPU architecture this is configured for

pub fn physical_regs(&self) -> &'static [PReg]

Returns a list of all the accessible registers for the architecture

pub fn instruction_pointer(&self) -> PReg

Gets the register that refers to the instruction pointer

pub fn options(&self) -> CodegenOptions

Gets the CodegenOptions that the code is being generated according to.