Trait sapphire::codegen::Architecture

pub trait Architecture {
    type Inst: MachInst<Arch = Self>;
    type Data: FuncData;

    // Required methods
    fn cpu() -> CPUArch;
    fn bool_layout() -> TypeLayout;
    fn ptr_layout() -> TypeLayout;
    fn i8_layout() -> TypeLayout;
    fn i16_layout() -> TypeLayout;
    fn i32_layout() -> TypeLayout;
    fn i64_layout() -> TypeLayout;
    fn f32_layout() -> TypeLayout;
    fn f64_layout() -> TypeLayout;
    fn physical_regs() -> &'static [PReg];
    fn instruction_pointer() -> PReg;
}

Models the architecture-specific details that a backend needs to deal with, mainly fundamental type layouts.

Required Associated Types

type Inst: MachInst<Arch = Self>

The MIR instruction type used for the architecture

type Data: FuncData

The type of function-associated data that this architecture has in its MIR

Required Methods

fn cpu() -> CPUArch

Returns the specific CPU architecture this is configured for

fn bool_layout() -> TypeLayout

Returns the layout of a bool

fn ptr_layout() -> TypeLayout

Returns the layout of a ptr

fn i8_layout() -> TypeLayout

Returns the layout of an i8

fn i16_layout() -> TypeLayout

Returns the layout of an i16

fn i32_layout() -> TypeLayout

Returns the layout of an i32

fn i64_layout() -> TypeLayout

Returns the layout of an i64

fn f32_layout() -> TypeLayout

Returns the layout of an f32

fn f64_layout() -> TypeLayout

Returns the layout of an f64

fn physical_regs() -> &'static [PReg]

Returns a list of all the accessible registers for the architecture

fn instruction_pointer() -> PReg

Gets the register that refers to the instruction pointer