Trait sapphire::codegen::CallingConv

pub trait CallingConv<Arch: Architecture> {
    // Required methods
    fn lower_call(
        &self,
        call: &CallInst,
        result: Option<WriteableReg>,
        context: Ctx<'_, '_, '_, '_, Arch>
    );
    fn lower_indirect_call(
        &self,
        call: &IndirectCallInst,
        result: Option<WriteableReg>,
        context: Ctx<'_, '_, '_, '_, Arch>
    );
}

Models a calling convention.

This handles emitting the code to properly call functions, including any stack alignment, parameter passing, etc.

Required Methods

fn lower_call( &self, call: &CallInst, result: Option<WriteableReg>, context: Ctx<'_, '_, '_, '_, Arch> )

Lowers a call instruction, handling passing arguments according to ABI rules. If the call returns a result, the result is stored into the register in result.

fn lower_indirect_call( &self, call: &IndirectCallInst, result: Option<WriteableReg>, context: Ctx<'_, '_, '_, '_, Arch> )

Lowers an indirect call instruction, handling passing arguments according to ABI rules. If the call returns a result, the result is stored into the register in result.