Trait sapphire::codegen::Emitter

pub trait Emitter<Arch: Architecture>: Sized {
    type AssemblyFormat;
    type ObjectCodeFormat;

    // Required methods
    fn assembly(
        module: &MIRModule<Arch::Inst>,
        format: Self::AssemblyFormat,
        target: TargetPair,
        fixed_interval_comments: bool
    ) -> String;
    fn object(
        module: &MIRModule<Arch::Inst>,
        format: Self::ObjectCodeFormat
    ) -> Vec<u8>;
}

An emitter that can emit different formats from a given type of MachInst.

This is divided into two main categories, assembly output and object code output.

Required Associated Types

type AssemblyFormat

Different types of assembly, if supported by the emitter. This should be used for e.g. different assembly dialects.

type ObjectCodeFormat

Different types of object code output, if supported by the emitter. This should be used for e.g. different file formats.

Required Methods

fn assembly( module: &MIRModule<Arch::Inst>, format: Self::AssemblyFormat, target: TargetPair, fixed_interval_comments: bool ) -> String

Emits assembly in a format specified by the emitter, returning a string that can be written to a file or printed.

Note that fixed_interval_comments only works when regalloc has not been done.

fn object( module: &MIRModule<Arch::Inst>, format: Self::ObjectCodeFormat ) -> Vec<u8>

Emits object code in a format specified by the emitter, returning a byte buffer containing the correctly-formatted object code that can be written to a file.