Trait sapphire::codegen::InstructionSelector

pub trait InstructionSelector<'module, 'target, Arch: Architecture>: Sized {
    // Required methods
    fn new() -> Self;
    fn frame_for_func(
        func: &'module Function,
        target: &Target<Arch>
    ) -> Box<dyn StackFrame<Arch>>;
    fn lower(
        &mut self,
        inst: Inst,
        def: &'module FunctionDefinition,
        frame: &mut dyn StackFrame<Arch>,
        ctx: &mut LoweringContext<'module, 'target, Arch>
    );
}

An instruction selector that lowers one Inst into (potentially many) MachInsts.

Specific implementations of this trait may be generic (e.g. working with multiple ABIs), but for any selector it will produce code that is ready to be register allocated.

Required Methods

fn new() -> Self

Creates an instance of the selector with itself configured correctly.

fn frame_for_func( func: &'module Function, target: &Target<Arch> ) -> Box<dyn StackFrame<Arch>>

If funcs calling convention is supported, returns a stack frame for that calling convention.

fn lower( &mut self, inst: Inst, def: &'module FunctionDefinition, frame: &mut dyn StackFrame<Arch>, ctx: &mut LoweringContext<'module, 'target, Arch> )

Lowers a single SIR instruction into MachInsts for the specified architecture.

This must report what it does to ctx in order for the isel pipeline to correctly lower a module. If any SIR instructions are merged when lowering the current instruction, this must be reported to ctx for correctness.