Trait sapphire::codegen::MachInst

pub trait MachInst: Copy + Debug + Hash + Sized {
    type Arch: Architecture;

    // Required methods
    fn uses<const N: usize>(
        &self,
        frame: &dyn StackFrame<Self::Arch>,
        collector: &mut RegCollector<N>
    );
    fn defs<const N: usize>(
        &self,
        frame: &dyn StackFrame<Self::Arch>,
        collector: &mut RegCollector<N>
    );
    fn as_copy(&self) -> Option<RegToRegCopy>;
    fn as_unconditional_jmp(&self) -> Option<UnconditionalBranch>;
    fn copy(width: usize, src: Reg, dest: WriteableReg) -> Self;
    fn load(width: usize, from: VariableLocation, to: PReg) -> Self;
    fn store(width: usize, from: PReg, to: VariableLocation) -> Self;
    fn is_ret(&self) -> bool;
    fn rewrite(
        self,
        rewrites: &[(Reg, PReg)],
        frame: &dyn StackFrame<Self::Arch>
    ) -> Self;

    // Provided methods
    fn is_copy(&self) -> bool { ... }
    fn is_nop_copy(&self) -> bool { ... }
    fn is_unconditional_jmp(&self) -> bool { ... }
}

Defines the API for a machine instruction.

Each specific target has a unique set of possible MachInsts (this is what makes up the unique “MIR” (machine instruction representation) for that target), but they all provide the same basic operations. This enables cross-target code sharing for things like instruction scheduling and register allocation.

Required Associated Types

type Arch: Architecture

The related architecture type that defines the architecture this specific MIR type targets

Required Methods

fn uses<const N: usize>( &self, frame: &dyn StackFrame<Self::Arch>, collector: &mut RegCollector<N> )

Pushes every register being used by self into collector.

It can be assumed that if the live ranges of these end at this instruction that they can be used as the destination.

fn defs<const N: usize>( &self, frame: &dyn StackFrame<Self::Arch>, collector: &mut RegCollector<N> )

Pushes every register being defined by self into collector.

fn as_copy(&self) -> Option<RegToRegCopy>

Returns self as a RegToRegCopy if self is a register-to-register copy.

fn as_unconditional_jmp(&self) -> Option<UnconditionalBranch>

Returns self as a UnconditionalBranch if self is an unconditional branch

fn copy(width: usize, src: Reg, dest: WriteableReg) -> Self

Creates a target-specific copy instruction that copies a value of width bytes between two registers.

This must return an object where Self::is_copy returns true.

fn load(width: usize, from: VariableLocation, to: PReg) -> Self

Creates an instruction that loads from into to.

fn store(width: usize, from: PReg, to: VariableLocation) -> Self

Creates an instruction that stores from into to.

fn is_ret(&self) -> bool

Checks if the instruction is a return instruction

fn rewrite( self, rewrites: &[(Reg, PReg)], frame: &dyn StackFrame<Self::Arch> ) -> Self

Rewrites an instruction to change all Reg references into PRegs.

Provided Methods

fn is_copy(&self) -> bool

Checks if the instruction is a copy between two registers.

fn is_nop_copy(&self) -> bool

Checks if the instruction is a copy from the same register being copied to.

These %r <- %r copies are useless and are able to be removed by the rewriter.

fn is_unconditional_jmp(&self) -> bool

Checks if the instruction is an unconditional jump