Trait sapphire::codegen::RegisterAllocator

pub trait RegisterAllocator<Arch: Architecture>: Sized {
    // Required method
    fn allocate(
        program: &MIRFunction<Arch::Inst>,
        frame: &mut dyn StackFrame<Arch>
    ) -> Allocation;
}

Models a register allocator that works over MIR.

The input to a register allocator is expected to be non-allocated MIR code (i.e. code that uses virtual registers, potentially with some pre-allocated operations that are bound to specific physical registers).

The output is a (representation of) a valid register allocation for the program that can be fed into the Rewriter, containing a list of spills/reloads to insert, and a mapping of virtual registers to physical registers..

Required Methods

fn allocate( program: &MIRFunction<Arch::Inst>, frame: &mut dyn StackFrame<Arch> ) -> Allocation

Computes a valid register allocation for the given “program” (function).

If the allocator performs any spills, it tells frame through StackFrame::spill_slot.