Trait sapphire::codegen::StackFrame

pub trait StackFrame<Arch: Architecture> {

    // Required methods
    fn stack_slot(
        &mut self,
        stack: StackSlot,
        context: FramelessCtx<'_, '_, '_, Arch>
    ) -> VariableLocation;
    fn lower_parameter(
        &mut self,
        param: Value,
        result: WriteableReg,
        context: FramelessCtx<'_, '_, '_, Arch>
    );
    fn lower_ret(
        &mut self,
        ret: &RetInst,
        context: FramelessCtx<'_, '_, '_, Arch>
    );
    fn spill_slot(&mut self, bytes: usize) -> VariableLocation;
    fn preserved_reg_used(&mut self, reg: PReg);
    fn register_use_def_call(
        &mut self,
        uses: &[PReg],
        defs: &[PReg]
    ) -> CallUseDefId;
    fn call_use_defs(&self, id: CallUseDefId) -> (&[PReg], &[PReg]);
    fn ret_uses(&self) -> &[PReg];
    fn generate_prologue(&self, out: &mut Vec<Arch::Inst>);
    fn generate_epilogue(&self, out: &mut Vec<Arch::Inst>);
    fn registers(&self) -> AvailableRegisters;
    fn metadata(&self) -> FunctionMetadata;
    fn stack_size(&self) -> usize;

}

Interface for a generic “stack frame” that a target can implement. This is used by the code generator to place data in the correct place at the function level.

This is a stateful type, the code generator and register allocator are both expected to use this when manipulating the stack and whatnot.

Required Methods

fn stack_slot( &mut self, stack: StackSlot, context: FramelessCtx<'_, '_, '_, Arch> ) -> VariableLocation

Informs the frame of a stack object that is being used. The frame returns a location that the stack object can be accessed through.

fn lower_parameter( &mut self, param: Value, result: WriteableReg, context: FramelessCtx<'_, '_, '_, Arch> )

Emits a copy from the location of a function parameter into a vreg during instruction selection.

fn lower_ret(&mut self, ret: &RetInst, context: FramelessCtx<'_, '_, '_, Arch>)

Lowers a return instruction, handling moving values into the correct spot according to the ABI.

fn spill_slot(&mut self, bytes: usize) -> VariableLocation

“Spills” a register that is bytes bytes long, returning a stack location for it.

It’s up to the caller to store this location and re-use it.

fn preserved_reg_used(&mut self, reg: PReg)

Used to inform a frame that a “callee-preserved” register has been used directly and thus needs to be preserved.

fn register_use_def_call( &mut self, uses: &[PReg], defs: &[PReg] ) -> CallUseDefId

Used to inform a frame of register use/def information for call instructions.

This is called by the calling convention when a call-like instruction is generated, with a list of used registers and a list of defined registers.

fn call_use_defs(&self, id: CallUseDefId) -> (&[PReg], &[PReg])

Returns the use-def information previously given to Self::register_use_def_call.

fn ret_uses(&self) -> &[PReg]

Returns the use-def information of a ret instruction

fn generate_prologue(&self, out: &mut Vec<Arch::Inst>)

Called during the final emitting phase, will emit a function’s prologue (if necessary).

This relies on the fact that the frame is maintained through all backend phases, so the object will know if a prologue is necessary.

fn generate_epilogue(&self, out: &mut Vec<Arch::Inst>)

Called during the final emitting phase, will emit a function’s epilogue (if necessary).

This relies on the fact that the frame is maintained through all backend phases, so the object will know if an epilogue is necessary.

fn registers(&self) -> AvailableRegisters

Returns a representation of the available registers for allocation.

fn metadata(&self) -> FunctionMetadata

Gets the metadata for the associated SIR function at the time of lowering

fn stack_size(&self) -> usize

The number of bytes that have been put on the stack at this point in the lowering