Module sapphire::codegen::x86_64

x86-64 Backend

This is the main module for code specific to the generic “x86-64 backend,” with code for general x86-64 codegen and code for specific ABIs on x86-64.

Structs

ALU

An ALU instruction with a given opcode that operates on two registers.

Call

Calls a known function

Cdq

Copies the sign bit of the 32-bit source (always eax) into the dest (always edx)

Cmp

A cmp instruction. Subtracts two values, discards the result, and updates CPU flags.

Cqo

Copies the sign bit of the 64-bit source (always rax) into the dest (always rdx)

Cwd

Copies the sign bit of the 16-bit source (always ax) into the dest (always dx)

Debug3RegLinuxX86_64

NOT INTENDED FOR PRODUCTION USE.

Debug3RegStackFrame

A debug stack frame setup that only makes 3 registers available for the register allocator.

Div

Performs unsigned integer division. The following assumptions are made:

Emit

The x86-64 emitter. This delegates to more specialized (and internal) mechanisms to do the actual emitting.

FrameInfo

Information about the stack frame of a function

GreedyISel

An instruction selector for x86-64.

IDiv

Performs signed integer division. The following assumptions are made:

IMul

Multiplies two integers and gets a 128bit result

IndirectCall

Performs an indirect call to a register, function pointer in memory, etc.

Jump

Possibly jumps to a block or target

Lea

A lea that computes an address, and then puts it into dest

LinuxX86_64

A platform representing (typical) Linux-based operating systems.

MacOSX86_64

A platform representing the (x86-64 variation of) the macOS operating systems.

Mov

A mov into a register

MovStore

A mov that stores into memory

Movabs

Copies a 64-bit value into a register

Movsx

Sign-extending copy instruction between two locations

Movzx

Zero-extending copy instruction between two locations

Neg

Negates the (signed) value in a register

Nop

A nop instruction with a specified byte length

Not

Performs bitwise NOT on a register

Pop

Pops a value from the stack and writes it to dest

Push

Pushes the value of a register onto the stack

Ret

Returns from the current function

ScaleAnd30BitOffset

This type is a compact representation of a (Scale, i32) tuple, with a slight limitation that the i32 must actually be representable in 2’s complement 30-bit (i.e. must be within the range [-2^29, 2^29 - 1]).

Set

A setCC instruction, used to set a byte based on a condition

SystemVCallingConv

Models the System-V calling convention

SystemVStackFrame

An implementation of StackFrame for the System-V ABI.

Test

A test instruction. Performs bitwise AND between two values, discards the result, and updates CPU flags.

Ud2

An undefined instruction that generates a trap when executed

WidthPair

Extremely compact way of storing the width (in bytes) of the input and output of an instruction.

WindowsX64CallingConv

An implementation of CallingConv for the Windows x64 ABI.

WindowsX64StackFrame

An implementation of StackFrame for the Windows x64 ABI.

WindowsX86_64

A platform representing the (x86-64 variation of) the Windows operating system.

X86_64

A target for a generic x86_64 CPU.

Enums

ALUOpcode

The opcode of A specific ALU instruction

ConditionCode

A condition code, used by several instructions to do operations based on the state of the CPU’s status flags

IndirectAddress

An operand to an instruction that references a memory location and loads a value of a specific size.

Inst

Models a single machine instruction for the x86_64 backend.

JumpTarget

A target that a jmp instruction could target

RegImm

The possible operands for instructions that can deal with either registers or in-line memory accesses

RegMem

The possible operands for instructions that can deal with either registers or in-line memory accesses

RegMemImm

The possible operands for instructions that can deal with more than just registers

Scale

The possible scalar (multiply) values for indirect addressing.

Width

The size of an operand. For Regs, this is the specific size of that register being used. For IndirectAddresss, this is the size of the value being loaded.

X86_64Assembly

Different assembly formats for x86-64

X86_64ObjectFile

Different object file formats for x86-64

Functions

format_reg

Formats the “canonical” assembly form of an x86-64 register