Enum sapphire::codegen::x86_64::Inst

#[repr(u32)]
pub enum Inst {

    Nop(Nop),
    Mov(Mov),
    Movzx(Movzx),
    Movsx(Movsx),
    MovStore(MovStore),
    Movabs(Movabs),
    Movaps(Movaps),
    MovFloatStore(MovFloatStore),
    MovFloatLoad(MovFloatLoad),
    Lea(Lea),
    ALU(ALU),
    Not(Not),
    Neg(Neg),
    IMul(IMul),
    Cwd(Cwd),
    Cdq(Cdq),
    Cqo(Cqo),
    Div(Div),
    IDiv(IDiv),
    PXor(PXor),
    FloatArith(FloatArith),
    Cmp(Cmp),
    Test(Test),
    Set(Set),
    Push(Push),
    Pop(Pop),
    Jump(Jump),
    Call(Call),
    IndirectCall(IndirectCall),
    Ret(Ret),
    Ud2(Ud2),

}

Models a single machine instruction for the x86_64 backend.

Variants

Nop(Nop)

A no-op instruction

Mov(Mov)

A mov into a register

Movzx(Movzx)

A movzx that performs zero-extension

Movsx(Movsx)

A movsx that performs sign-extension

MovStore(MovStore)

A mov that stores into memory

Movabs(Movabs)

Moves a 64-bit constant into a register

Movaps(Movaps)

Moves floating-point values between registers

MovFloatStore(MovFloatStore)

Stores a floating-point value into memory

MovFloatLoad(MovFloatLoad)

Loads a floating-point value from memory

Lea(Lea)

An lea instruction that computes an address

ALU(ALU)

An ALU instruction with a given opcode that operates on two registers.

Not(Not)

Performs bitwise NOT on a register

Neg(Neg)

Negates the (signed) value in a register

IMul(IMul)

A multiplication that produces a 128-bit result

Cwd(Cwd)

Broadcasts the sign bit of ax into dx

Cdq(Cdq)

Broadcasts the sign bit of eax into edx

Cqo(Cqo)

Broadcasts the sign bit of rax into rdx

Div(Div)

Performs unsigned division

IDiv(IDiv)

Performs signed division

PXor(PXor)

Floating-point bitwise XOR

FloatArith(FloatArith)

Floating-point arithmetic instruction

Cmp(Cmp)

A comparison between two integers

Test(Test)

A different comparison between two integers

Set(Set)

Sets a byte based on a condition

Push(Push)

Pushes the value of a register onto the stack

Pop(Pop)

Pops a value from the stack and writes it to dest

Jump(Jump)

A jump to a local location in the function

Call(Call)

Calls a known function

IndirectCall(IndirectCall)

Performs an indirect call

Ret(Ret)

Returns from the function

Ud2(Ud2)

An undefined instruction that generates a trap when executed