Struct sapphire::codegen::x86_64::Debug3RegStackFrame

pub struct Debug3RegStackFrame { /* private fields */ }

A debug stack frame setup that only makes 3 registers available for the register allocator.

Implementations

impl Debug3RegStackFrame

pub fn wrap(frame: SystemVStackFrame) -> Self

Creates a new debug stack frame that wraps frame]