Enum sapphire::ir::InstData

#[repr(u32)]
pub enum InstData {

    Call(CallInst),
    IndirectCall(IndirectCallInst),
    ICmp(ICmpInst),
    FCmp(FCmpInst),
    Sel(SelInst),
    Br(BrInst),
    CondBr(CondBrInst),
    Unreachable(UnreachableInst),
    Ret(RetInst),
    And(CommutativeArithInst),
    Or(CommutativeArithInst),
    Xor(CommutativeArithInst),
    Shl(ArithInst),
    AShr(ArithInst),
    LShr(ArithInst),
    IAdd(CommutativeArithInst),
    ISub(ArithInst),
    IMul(CommutativeArithInst),
    SDiv(ArithInst),
    UDiv(ArithInst),
    SRem(ArithInst),
    URem(ArithInst),
    FNeg(FloatUnaryInst),
    FAdd(ArithInst),
    FSub(ArithInst),
    FMul(ArithInst),
    FDiv(ArithInst),
    FRem(ArithInst),
    Alloca(AllocaInst),
    Load(LoadInst),
    Store(StoreInst),
    Offset(OffsetInst),
    Extract(ExtractInst),
    Insert(InsertInst),
    ElemPtr(ElemPtrInst),
    Sext(CastInst),
    Zext(CastInst),
    Trunc(CastInst),
    IToB(CastInst),
    BToI(CastInst),
    SIToF(CastInst),
    UIToF(CastInst),
    FToSI(CastInst),
    FToUI(CastInst),
    FExt(CastInst),
    FTrunc(CastInst),
    IToP(CastInst),
    PToI(CastInst),
    IConst(IConstInst),
    FConst(FConstInst),
    BConst(BConstInst),
    Undef(UndefConstInst),
    Null(NullConstInst),
    GlobalAddr(GlobalAddrInst),

}

This holds both the opcode of a given instruction and all the state that makes up that specific instruction.

While each instruction may have wildly different actual data, they all are stored in the same table and all inside the same InstructionData type.

Variants

Call(CallInst)

call T @function(args...), models a direct call to a known function.

IndirectCall(IndirectCallInst)

call T %var(args...), models an indirect call through a pointer.

ICmp(ICmpInst)

icmp op T %a, %b, models an integer comparison

FCmp(FCmpInst)

fcmp op T %a, %b, models a floating-point comparison

Sel(SelInst)

sel bool %cond, T %a, T %b, models a ternary-like instruction

Br(BrInst)

br block, models an unconditional branch

CondBr(CondBrInst)

condbr bool %cond, if block1, else block2, models a conditional branch between two blocks

Unreachable(UnreachableInst)

unreachable, a terminator that can never be executed

Ret(RetInst)

ret T %val, returns from the current function

And(CommutativeArithInst)

and T %a, %b, performs bitwise AND

Or(CommutativeArithInst)

or T %a, %b, performs bitwise OR

Xor(CommutativeArithInst)

xor T %a, %b, performs bitwise XOR

Shl(ArithInst)

shl T %a, %b, performs bitwise left-shift

AShr(ArithInst)

ashr T %a, %b, performs bitwise arithmetic right-shift

LShr(ArithInst)

lshr T %a, %b, performs bitwise logical right-shift

IAdd(CommutativeArithInst)

iadd T %a, %b, performs two’s complement addition

ISub(ArithInst)

isub T %a, %b, performs two’s complement subtraction

IMul(CommutativeArithInst)

imul T %a, %b, performs two’s complement multiplication

SDiv(ArithInst)

sdiv T %a, %b, performs signed division

UDiv(ArithInst)

udiv T %a, %b, performs unsigned division

SRem(ArithInst)

srem T %a, %b, gets the remainder of performing signed division

URem(ArithInst)

urem T %a, %b, gets the remainder of performing unsigned division

FNeg(FloatUnaryInst)

fneg T %a, performs floating-point negation

FAdd(ArithInst)

fadd T %a, %b, performs floating-point addition

FSub(ArithInst)

fadd T %a, %b, performs floating-point subtraction

FMul(ArithInst)

fadd T %a, %b, performs floating-point multiplication

FDiv(ArithInst)

fadd T %a, %b, performs floating-point division

FRem(ArithInst)

fadd T %a, %b, gets the remainder of performing floating-point division

Alloca(AllocaInst)

alloca T, performs stack allocation

Load(LoadInst)

load T, ptr %p, loads from a pointer

Store(StoreInst)

store T %a, ptr %p, stores a value to a pointer

Offset(OffsetInst)

offset T, ptr %p, performs pointer arithmetic

Extract(ExtractInst)

extract T %s, N, extracts a value from an aggregate

Insert(InsertInst)

insert T %s, U %a, N, inserts a value into an aggregate

ElemPtr(ElemPtrInst)

elemptr T, ptr %p, N, gets a pointer into an aggregate

Sext(CastInst)

sext T, U %b, performs sign-extension

Zext(CastInst)

zext T, U %b, performs zero-extension

Trunc(CastInst)

trunc T, U %b, performs truncation

IToB(CastInst)

itob T, U %b, converts an integer into a bool

BToI(CastInst)

btoi T, U %b, converts a bool into an integer

SIToF(CastInst)

sitof T, U %b, converts a signed integer into a float

UIToF(CastInst)

itob T, U %b, converts an unsigned integer into a float

FToSI(CastInst)

ftosi T, U %b, converts a float into a signed integer

FToUI(CastInst)

ftoui T, U %b, converts a float into an unsigned integer

FExt(CastInst)

fext T, U %b, converts a float into a larger float

FTrunc(CastInst)

ftrunc T, U %b, converts a float into a smaller float

IToP(CastInst)

itop T, U %b, converts an integer into a pointer

PToI(CastInst)

ptoi T, U %b, converts a pointer into an integer

IConst(IConstInst)

iconst T N, materializes an integer constant

FConst(FConstInst)

fconst T N, materializes a floating-point constant

BConst(BConstInst)

bconst N, materializes a bool constant

Undef(UndefConstInst)

undef T N, materializes an uninitialized value

Null(NullConstInst)

null T N, materializes a null value

GlobalAddr(GlobalAddrInst)

globaladdr @name, materializes a pointer to a global value

Implementations

impl InstData

pub fn opc(&self) -> Discriminant<InstData>

Gets the discriminant of the InstData, this is the “opcode” of the instruction. This can be used to trivially check if two instructions are the same variant without needing the mem::discriminant boilerplate.