Module sapphire::ir::matchers

Defines APIs for pattern-matching on SIR.

This is meant for use at the SIR level, instruction selection needs different matchers due to extra invariants needing to be checked.

let mut module = Module::new("test");
let mut b = module.define_function("test", SigBuilder::new().ret(Some(Type::i32())).build());
let bb0 = b.create_block("bb0");
b.switch_to(bb0);

let v1 = b.append().iconst(Type::i32(), 42, DebugInfo::fake());
let v2 = b.append().iconst(Type::i32(), 16, DebugInfo::fake());
let v3 = b.append().iadd(v2, v1, DebugInfo::fake());
let v4 = b.append().imul(v1, v3, DebugInfo::fake());
b.append().ret_val(v4, DebugInfo::fake());

let f = b.define();
let dfg = &module.function(f).definition().unwrap().dfg;

let iconst_42 = iconst_val(42);
let iconst_i32 = iconst_ty(Type::i32());
let iadd_i32_42 = iadd_with(val_of_ty(Type::i32()), iconst_ty_val(Type::i32(), 42));

assert!(matches(v1, iconst_42, dfg));
assert!(matches(v3, iadd_i32_42, dfg));

Structs

BasicInstMatcher

A matcher that wraps up a matcher function.

BasicValMatcher

A matcher that matches on values

Traits

IRMatcher

A basic matcher for a single value/instruction in the IR.

Functions

and

Allows matching against and instructions.

and_with

Allows matching against and instructions and their lhs/rhs.

ashr

Allows matching against ashr instructions.

ashr_with

Allows matching against ashr instructions and their lhs/rhs.

bconst

A matcher that matches on bconst instructions (i.e. matches on constant operands).

bconst_val

A matcher that matches on bconst instructions with a specific value

binary

A matcher that matches any binary instruction.

binary_with

A matcher that matches a binary instruction.

block_param

A matcher that matches any value coming from a φ node.

block_param_of_ty

A matcher that matches any value coming from a φ node that also has a given type.

both

Logical conjunction operation between two matchers.

fadd

Allows matching against fadd instructions.

fadd_with

Allows matching against fadd instructions and their lhs/rhs.

fcmp

Allows matching against fcmp instructions.

fcmp_with

Allows matching against fcmp instructions and their lhs/rhs.

fconst

A matcher that matches on fconst instructions (i.e. matches on constant operands).

fconst_ty

A matcher that matches on fconst instructions with a specific type

fdiv

Allows matching against fdiv instructions.

fdiv_with

Allows matching against fdiv instructions and their lhs/rhs.

fmul

Allows matching against fmul instructions.

fmul_with

Allows matching against fmul instructions and their lhs/rhs.

frem

Allows matching against frem instructions.

frem_with

Allows matching against frem instructions and their lhs/rhs.

fsub

Allows matching against fsub instructions.

fsub_with

Allows matching against fsub instructions and their lhs/rhs.

iadd

Allows matching against iadd instructions.

iadd_with

Allows matching against iadd instructions and their lhs/rhs.

icmp

Allows matching against icmp instructions.

icmp_with

Allows matching against icmp instructions and their lhs/rhs.

iconst

A matcher that matches on iconst instructions (i.e. matches on constant operands).

iconst_ty

A matcher that matches on iconst instructions with a specific type

iconst_ty_val

A matcher that matches on iconst instructions with a specific value and a specific type.

iconst_val

A matcher that matches on iconst instructions with a specific value

imul

Allows matching against imul instructions.

imul_with

Allows matching against imul instructions and their lhs/rhs.

isub

Allows matching against isub instructions.

isub_with

Allows matching against isub instructions and their lhs/rhs.

load

Matches a load instruction

lshr

Allows matching against lshr instructions.

lshr_with

Allows matching against lshr instructions and their lhs/rhs.

matches

Returns whether or not a value matches the provided matcher

matches_inst

Returns whether or not an instruction matches the provided matcher

null

A matcher that matches on null instructions (i.e. matches on constant operands).

null_float

A matcher that matches on null instructions with a floating-point type

null_int

A matcher that matches on null instructions with an integer type

null_integral

A matcher that matches on null instructions with an integer, boolean or pointer type

null_ty

A matcher that matches on null instructions with a specific type

one_of

Logical disjunction operation between two matchers.

or

Allows matching against or instructions.

or_with

Allows matching against or instructions and their lhs/rhs.

power_of_two

Matches on an iconst value that is a power of 2

sdiv

Allows matching against sdiv instructions.

sdiv_with

Allows matching against sdiv instructions and their lhs/rhs.

shl

Allows matching against shl instructions.

shl_with

Allows matching against shl instructions and their lhs/rhs.

srem

Allows matching against srem instructions.

srem_with

Allows matching against srem instructions and their lhs/rhs.

stackslot

Matches a stackslot instruction

udiv

Allows matching against udiv instructions.

udiv_with

Allows matching against udiv instructions and their lhs/rhs.

undef

A matcher that matches on undef instructions (i.e. matches on constant operands).

urem

Allows matching against urem instructions.

urem_with

Allows matching against urem instructions and their lhs/rhs.

val

A matcher that matches any value

val_of_ty

A matcher that matches any value with a given type

xor

Allows matching against xor instructions.

xor_with

Allows matching against xor instructions and their lhs/rhs.