Struct sapphire::ir::DataFlowGraph

pub struct DataFlowGraph { /* private fields */ }

Owns all of the instructions, basic blocks, values, and everything else in a given function. Also models all the complex data-flow information between various instructions, although it does not model any of the actual code layout information (block ordering, instruction ordering, etc).

Implementations

impl DataFlowGraph

pub fn new() -> Self

Creates a new data-flow graph for a new function.

pub fn signature(&self, sig: Sig) -> &Signature

Gets a function’s Signature from a given Sig. Any Sig used by any indirect or direct calls inside the function body can be resolved here.

pub fn value_to_inst(&self, value: Value) -> Option<Inst>

Converts a Value into an Inst if and only if that value refers to an instruction’s result. If value refers to a block parameter instead, None is returned.

pub fn inst_to_result(&self, inst: Inst) -> Option<Value>

Converts an Inst into a Value referring to its result if and only if that instruction actually yields a result. If it does not have a result, None is returned.

pub fn data(&self, inst: Inst) -> &InstData

Gets a single instruction’s InstData.

pub fn inst_debug(&self, inst: Inst) -> DebugInfo

Gets a single instruction’s DebugInfo.

pub fn debug(&self, inst: Value) -> DebugInfo

Gets a single value’s DebugInfo.

pub fn ty(&self, value: Value) -> Type

Gets the type of the value that a given Value evaluates to.

pub fn operands(&self, value: Value) -> &[Value]

Gets any values that are used as operands for computing value.

This can potentially be empty, e.g. for block params or constant materialization instructions.

pub fn create_inst( &mut self, data: InstData, debug: DebugInfo ) -> (Inst, Option<Value>)

Inserts an instruction into the DFG, and returns a reference to it. If the instruction yields a result (and thus can also be used as an operand for other instructions), that value is also returned as the second return value.

pub fn create_block(&mut self, name: Str) -> Block

Inserts a basic block with a given name into the DFG. It will start with an empty list of block parameters, these can be appended later.

pub fn find_block(&self, name: Str) -> Option<Block>

Finds a block by name, if it exists.

pub fn block(&self, block: Block) -> &BasicBlock

Resolves a block into a full BasicBlock.

pub fn remove_block(&mut self, block: Block)

Removes a basic block that already exists.

pub fn append_block_param( &mut self, bb: Block, ty: Type, debug: DebugInfo ) -> Value

Appends a block parameter to a given block.

pub fn insert_sig(&mut self, sig: &Signature) -> Sig

Inserts a function signature into the signature arena (if it isn’t already in the arena), and returns a Sig that can refer to it.

pub fn is_inst_inserted(&self, inst: Inst) -> bool

Checks if the DFG contains a given instruction. If it does not, the instruction is invalid. Any instruction returned from one of the insert_inst_* methods will be contained.

pub fn is_val_inserted(&self, value: Value) -> bool

Checks if the DFG contains a given value. If it does not, the value is invalid. Any values returned from one of the insert_inst_* methods or the Self::append_block_param method will be contained.

pub fn is_block_inserted(&self, block: Block) -> bool

Checks if the DFG contains a given block. If the block was returned from any of the block insertion methods, it will be contained.

pub fn value_def(&self, val: Value) -> ValueDef

Gets the definition of a given value

pub fn is_block_param(&self, val: Value) -> bool

Checks if a given value is a block parameter.

pub fn branch_info(&self, inst: Inst) -> Option<&[BlockWithParams]>

If an instruction is a terminator, returns the list of branch targets. If it isn’t, returns None.

pub fn replace_inst( &mut self, inst: Inst, data: InstData, debug: DebugInfo ) -> (Inst, Option<Value>)

Replaces inst with new data and new debuginfo. This makes all references to that Inst point to this new instruction instead.

pub fn uses_of(&self, val: Value) -> &[Inst]

Returns every use in the data-flow graph of a given value.

Note that all of these may not actually exist in the function layout. If you don’t want to pointlessly iterate over these, filter this by the ones that are present in a given layout.

pub fn replace_uses_with(&mut self, original: Value, new: Value)

Replaces every use of original with new.

pub fn next_value(&self) -> Value

Gets one past the highest possible value key. This is an invalid key, but it is suitable for use with SecondaryMap::fill.

pub fn remove_block_param(&mut self, block: Block, param: Value)

Removes a block parameter from a given block without invalidating references to the other block parameters. Note that this does change the index for other block parameters, but it doesn’t make their Values invalid.

pub fn rewrite_branch_args( &mut self, branch: Inst, target: Block, new_params: &[Value] )

Rewrites a branch instruction that targets target to have a new set of block params

pub fn replace_branch_arg(&mut self, inst: Inst, to: Block, arg: usize, new: Value)

Replaces a single argument to a branch target on a branch instruction. This is intended to be more efficient for the more common case where a value is only being replaced instead of inserted/removed.

pub fn create_stack_slot(&mut self, name: Str, ty: Type) -> StackSlot

Creates a new stack slot with a given name and type.

pub fn stack_slot(&self, slot: StackSlot) -> StackSlotData

Gets the information about a particular stack slot

pub fn stack_slots(&self) -> impl Iterator<Item = StackSlot> + '_

Provides an iterator over every stack slot in the function

pub fn is_stack_slot_inserted(&self, slot: StackSlot) -> bool

Checks if a given slot is actually a valid stack slot

pub fn remove_stack_slot(&mut self, slot: StackSlot)

Removes a stack slot that already exists.