Struct sapphire::ir::FuncBuilder

pub struct FuncBuilder<'m> { /* private fields */ }

Helper type for building a SIR function.

Implementations

impl<'m> FuncBuilder<'m>

pub fn define(self) -> Func

Finishes defining the function, and actually generates a full definition and inserts it into the module. Until this method is called, the function is not actually defined in the module.

pub fn find_block(&mut self, name: &str) -> Option<Block>

Finds a block by its name, returning it if it’s inserted into the current function.

pub fn block_name(&self, block: Block) -> &str

Gets the name of a block that has been inserted into the function

pub fn block_params(&self, block: Block) -> &[Value]

Gets the block parameters of a given block.

pub fn create_block(&mut self, name: &str) -> Block

Creates a single basic block and returns it. This block is appended to the end of the block list.

Note that this does not switch the builder to operate on that block, you still need to call Self::switch_to.

pub fn create_block_before(&mut self, name: &str, before: Block) -> Block

Equivalent to Self::create_block, except it inserts the block before before instead of appending it.

pub fn create_block_after(&mut self, name: &str, after: Block) -> Block

Equivalent to Self::create_block, except it inserts the block after after instead of appending it.

pub fn switch_to(&mut self, block: Block)

Switches to inserting at a specific block.

pub fn append_block_param(
&mut self,
block: Block,
ty: Type,
debug: DebugInfo
) -> Value

Adds a single block parameter of ty to block and returns a value that refers to it.

pub fn append_entry_params(
&mut self,
block: Block,
debug: DebugInfo
) -> TinyArray<Value, 2>

Appends all the block parameters necessary for the parameters of the current function. This is meant to be used for creating a function’s entry block, as the block parameters to the entry block get their value from the parameters.

This is equivalent to calling Self::append_block_param once for each parameter type in the function signature.

pub fn ty(&self, value: Value) -> Type

Gets the type of a value that was previously emitted by the builder.

pub fn append(&mut self) -> AppendBuilder<'_>

Returns a builder that can be used to append an instruction to the current block.

If there is no current block, trying to build an instruction will panic somewhere in the internals.

pub fn import_signature(&mut self, signature: &Signature) -> Sig

Imports a signature into the function and returns a Sig that refers to it.

This is intended to be used for calling other functions, where the signature needs to be known inside of the call instruction

pub fn signature(&self, sig: Sig) -> &Signature

Resolves a Sig into a real signature. This signature must have been yielded at some point from Self::import_signature.

pub fn inst_to_result(&self, inst: Inst) -> Option<Value>

Converts an Inst into a Value that refers to the result of the instruction if possible.

Not all instructions actually yield results, those will return None

pub fn type_pool(&self) -> &TypePool

Allows access to the TypePool of the module that the builder is borrowing.

pub fn type_pool_mut(&mut self) -> &mut TypePool

Allows mutable access to the TypePool of the module that the builder is borrowing.

pub fn string_pool(&self) -> &StringPool

Allows access to the TypePool of the module that the builder is borrowing.

pub fn string_pool_mut(&mut self) -> &mut StringPool

Allows mutable access to the TypePool of the module that the builder is borrowing.

pub fn function(&self, func: Func) -> &Function

Resolves a Func into a real function object.

pub fn function_mut(&mut self, func: Func) -> &mut Function

Resolves a Func into a real function object.

pub fn find_function_by_name(&self, func: &str) -> Option<Func>

Finds a Func with a given name. If the function has not been added to the module, None is returned.

pub fn is_entry_block(&self, block: Block) -> bool

Checks if a given block is the entry block to the function

pub fn entry_block(&self) -> Option<Block>

Gets the entry block of the function. Unless no blocks have been appended to the function, this will be Some.

pub fn current_func(&self) -> Func

Gets a Func referring to the function being built.

pub fn current_signature(&self) -> &Signature

Gets the Signature of the function being built.