Struct sapphire::ir::FuncCursor

pub struct FuncCursor<'f> {
    pub func: &'f mut Function,
    /* private fields */
}

Similar to FuncBuilder but for in-place modification of functions.

Fields
func: &'f mut Function

The function being moved over, public so it can be re-borrowed.

Implementations

impl<'f> FuncCursor<'f>

pub fn over(func: &'f mut Function) -> Self

Creates a FuncCursor that allows the given function to be modified.