Struct sapphire::ir::FuncView

pub struct FuncView<'f> { /* private fields */ }

Effectively a FuncCursor without any of the operations that mutate the function.

Implementations

impl<'f> FuncView<'f>

pub fn over(func: &'f Function) -> Self

Creates a FuncView that allows the given function to be viewed.