Struct sapphire::ir::Layout

pub struct Layout { /* private fields */ }

Models the layout of an entire function and every basic-block in it.

Each block is modeled as a linked list to allow easy splicing and in-place editing, and the list of blocks is also modeled as a linked list for similar reasons.

Implementations

impl Layout

pub fn new() -> Self

Creates a new, empty layout that is linked to the given data-flow graph.

pub fn append_inst(&mut self, inst: Inst, block: Block)

Appends an instruction to the end of the specified block. If this is the final instruction being appended to the block, it must be a terminator instruction.

pub fn insert_inst_before(&mut self, inst: Inst, before: Inst)

Inserts inst into the same block as before, but directly before before.

pub fn insert_inst_after(&mut self, inst: Inst, after: Inst)

Inserts inst into the same block as after, but directly after after.

pub fn remove_inst(&mut self, inst: Inst)

Removes an instruction from the layout. It is expected that the instruction exists, because removing a non-existent instruction is almost certainly a bug.

pub fn remove_inst_if_exists(&mut self, inst: Inst) -> bool

In the case that removing an instruction that is possibly not in the layout is necessary, use this instead of Self::remove_inst.

Returns true an instruction was removed.

pub fn append_block(&mut self, block: Block)

Appends a block to the layout, putting it at the end of the list of blocks.

pub fn insert_block_before(&mut self, block: Block, before: Block)

Inserts a block before another block in the list.

pub fn insert_block_after(&mut self, block: Block, after: Block)

Inserts a basic block after another basic block.

pub fn len_blocks(&self) -> usize

Returns the number of blocks in the layout

pub fn len_insts(&self) -> usize

Returns the number of instructions in the layout

pub fn is_block_inserted(&self, block: Block) -> bool

Checks if a block is currently inside the layout

pub fn is_inst_inserted(&self, inst: Inst) -> bool

Checks if an instruction is currently inside the layout

pub fn blocks(&self) -> BlockIter<'_>

Gets an iterator over the blocks of the layout.

pub fn insts_in_block(&self, block: Block) -> InstIter<'_>

Gets an iterator over every instruction in a given block.

pub fn entry_block(&self) -> Option<Block>

Gets the entry block for the layout, if it exists.

pub fn block_next(&self, block: Block) -> Option<Block>

Gets the block that comes after block

pub fn block_prev(&self, block: Block) -> Option<Block>

Gets the block that comes before block

pub fn block_first_inst(&self, block: Block) -> Option<Inst>

Gets the first instruction in block

pub fn block_last_inst(&self, block: Block) -> Option<Inst>

Gets the last instruction in block

pub fn inst_next(&self, inst: Inst) -> Option<Inst>

Gets the instruction that comes after inst

pub fn inst_prev(&self, inst: Inst) -> Option<Inst>

Gets the instruction that comes before inst

pub fn inst_block(&self, inst: Inst) -> Block

Gets the block that an instruction is in

pub fn loc_compare(&self, a: Inst, b: Inst) -> Ordering

Compares two instructions that are from the same block, and returns their relative order compared to each-other.

Ordering::Less => a < b order-wise, i.e. b comes after a. Ordering::Equal => a == b. Ordering::Greater => a > b order-wise, i.e. a comes after b.