Struct sapphire::ir::Module

pub struct Module { /* private fields */ }

Contains all the data necessary for a single module of SIR.

This models all of the information that would be represented inside of a textual module of SIR.

Implementations

impl Module

pub fn new(name: &str) -> Self

Creates a new, empty module with a given name.

This name is expected to be unique across different Modules, but it doesn’t have to be for the correctness of any transforms.

pub fn function(&self, func: Func) -> &Function

Resolves a Func into a real function object.

pub fn function_mut(&mut self, func: Func) -> &mut Function

Resolves a Func into a real function object.

pub fn find_function_by_name(&self, func: &str) -> Option<Func>

Finds a Func with a given name. If the function has not been added to the module, None is returned.

pub fn define_function(&mut self, name: &str, sig: Signature) -> FuncBuilder<'_>

Declares and then defines a new function.

pub fn declare_function(&mut self, name: &str, sig: Signature) -> Func

Declares a function without providing it a definition. It can be defined later with Self::define_existing_function, or it can be left as-is if the function is opaque.

pub fn define_existing_function(&mut self, func: Func) -> FuncBuilder<'_>

Returns a FuncBuilder that will create a body for a previously-declared function.

If the function already has a body, using the builder will completely replace the previous body.

pub fn functions(&self) -> impl Iterator<Item = Func>

Returns an iterator over all of the functions in the module.

pub fn type_pool(&self) -> &TypePool

Gets the TypePool in use for all the functions inside of a module.

pub fn type_pool_mut(&mut self) -> &mut TypePool

Gets the TypePool in use for all the functions inside of a module.

pub fn string_pool(&self) -> &StringPool

Gets the TypePool in use for all the functions inside of a module.

pub fn string_pool_mut(&mut self) -> &mut StringPool

Gets the TypePool in use for all the functions inside of a module.

pub fn insert_string(&mut self, string: &str) -> Str

Convenience method for self.string_pool_mut().insert(str).

pub fn resolve_string(&self, str: Str) -> &str

Convenience method for self.string_pool_mut()[str].

pub fn name(&self) -> &str

Gets the user-given name of the module.

pub fn identity(&self) -> ModuleIdentity

Gets a ModuleIdentity that refers to the object.