Struct sapphire::ir::SigBuilder

pub struct SigBuilder { /* private fields */ }

Helper type for building a Signature.

Implementations

impl SigBuilder

pub fn new() -> Self

Creates a SigBuilder for the signature void ()

pub fn vararg(self, value: bool) -> Self

Marks the signature as having a variable number of arguments.

pub fn abi(self, abi: CallConv) -> Self

Marks the function as having a specified ABI.

pub fn ret(self, ret: Option<Type>) -> Self

Marks the signature as having a given return type.

pub fn ret_with_attribute(self, ret: Option<(Type, RetAttributes)>) -> Self

Adds an attribute to the return value of the function

pub fn param(self, param: Type) -> Self

Appends a parameter to the signature

pub fn param_with_attribute( self, param: Type, attributes: ParamAttributes ) -> Self

Adds a param with an attribute list to the signature

pub fn params(self, params: &[Type]) -> Self

Appends a list of parameters to the signature

pub fn params_with_attributes(self, params: &[(Type, ParamAttributes)]) -> Self

Appends a list of parameters (all with attributes) to the signature

pub fn build(self) -> Signature

Builds the signature