Struct sapphire::ir::Type

pub struct Type { /* private fields */ }

A reference to a type. Copyable, compact, lightweight, and can model arbitrary types in the IR.

This type either contains all the information about the type (in the case of fundamental types), or contains a reference into a TypePool that holds the information about the type (only for arrays and structures).

let t1 = Type::bool(); // models `bool`
let t2 = Type::ptr(); // models `ptr`
assert_ne!(t1, t2);

For better pattern-matching it must be unpacked into an UType that has an actual enum representation, note that UType is the size of (u32, u32) instead of u32.

Implementations

impl Type

pub const fn bool() -> Self

Creates a boolean type (the bool type in the IR).

pub const fn ptr() -> Self

Creates a pointer type (the ptr type in the IR).

pub const fn int(width: u32) -> Self

Creates an integer type (the iN types in the IR) with a given width. Given width is $$N$$, $$N \in {8,16,32,64}$$ must be true.

let t1 = Type::int(32);
let t2 = Type::int(64);
assert_ne!(t1, t2);

pub const fn i8() -> Self

Shorthand for creating an integer of width 8. Exactly equivalent to Type::int(8).

let t1 = Type::i8();
let t2 = Type::int(8);

assert_eq!(t1, t2);

pub const fn i16() -> Self

Shorthand for creating an integer of width 16. Exactly equivalent to Type::int(16).

let t1 = Type::i16();
let t2 = Type::int(16);

assert_eq!(t1, t2);

pub const fn i32() -> Self

Shorthand for creating an integer of width 32. Exactly equivalent to Type::int(32).

let t1 = Type::i32();
let t2 = Type::int(32);

assert_eq!(t1, t2);

pub const fn i64() -> Self

Shorthand for creating an integer of width 64. Exactly equivalent to Type::int(64).

let t1 = Type::i64();
let t2 = Type::int(64);

assert_eq!(t1, t2);

pub const fn float(format: FloatFormat) -> Self

Creates a float type (the fN types in the IR), with the float being of a given format.

let t1 = Type::float(FloatFormat::Double);
let t2 = Type::float(FloatFormat::Single);
assert_ne!(t1, t2);

pub const fn f32() -> Self

Shorthand for Type::float(FloatFormat::Single). Exactly equivalent to that call.

let t1 = Type::f32();
let t2 = Type::float(FloatFormat::Single);
assert_eq!(t1, t2);

pub const fn f64() -> Self

Shorthand for Type::float(FloatFormat::Double). Exactly equivalent to that call.

let t1 = Type::f64();
let t2 = Type::float(FloatFormat::Double);
assert_eq!(t1, t2);

pub fn array(ctx: &mut TypePool, inner: Type, length: u64) -> Self

Creates an array type on the given context and returns it.

let mut ctx = TypePool::new();
let ty = Type::array(&mut ctx, Type::i32(), 16); // => [i32; 16]

pub fn structure(ctx: &mut TypePool, fields: &[Type]) -> Self

Creates a structure type from a given list of fields.

let mut ctx = TypePool::new();
let ty = Type::structure(&mut ctx, &[Type::i32(), Type::i32()]); // => { i32, i32 }

pub const fn is_bool(&self) -> bool

Checks if the type is a Bool.

let t1 = Type::bool();
assert_eq!(t1.is_bool(), true);

pub const fn is_ptr(&self) -> bool

Checks if the type is a Ptr.

let t1 = Type::ptr();
assert_eq!(t1.is_ptr(), true);

pub const fn is_int(&self) -> bool

Checks if the type is a Int.

let t1 = Type::int(32);
assert_eq!(t1.is_int(), true);

pub const fn is_int_of_width(&self, width: u32) -> bool

Checks if the type is a Int and has a given width.

let t1 = Type::int(64);
assert_eq!(t1.is_int_of_width(64), true);

pub const fn is_bool_or_int(&self) -> bool

Checks if a type is a bool or an iN.

let t1 = Type::int(64);
assert_eq!(t1.is_bool_or_int(), true);

pub const fn is_i8(&self) -> bool

Checks if the type is an Int with a width of 8.

let t1 = Type::i8();
assert_eq!(t1.is_i8(), true);

let t2 = Type::f64();
assert_eq!(t2.is_i8(), false);

pub const fn is_i16(&self) -> bool

Checks if the type is an Int with a width of 16.

let t1 = Type::i16();
assert_eq!(t1.is_i16(), true);

let t2 = Type::i8();
assert_eq!(t2.is_i16(), false);

pub const fn is_i32(&self) -> bool

Checks if the type is an Int with a width of 32.

let t1 = Type::i32();
assert_eq!(t1.is_i32(), true);

let t2 = Type::ptr();
assert_eq!(t2.is_i32(), false);

pub const fn is_i64(&self) -> bool

Checks if the type is an Int with a width of 64.

let t1 = Type::i64().unwrap_int();
assert_eq!(t1.is_i64(), true);

let t2 = Type::bool();
assert_eq!(t2.is_i64(), false);

pub const fn is_float(&self) -> bool

Checks if the type is a Float.

let t1 = Type::bool();
assert_eq!(t1.is_bool(), true);

pub const fn is_float_of_format(&self, format: FloatFormat) -> bool

Checks if the type is a Float and has a given format.

let t1 = Type::float(FloatFormat::Double);
assert_eq!(t1.is_float_of_format(FloatFormat::Double), true);

pub const fn is_f32(&self) -> bool

Checks if the type is a Float with format FloatFormat::Single.

let t1 = Type::f32();
assert_eq!(t1.is_f32(), true);

let t2 = Type::i32();
assert_eq!(t2.is_f32(), false);

pub const fn is_f64(&self) -> bool

Checks if the type is a Float with format FloatFormat::Double.

let t1 = Type::f64();
assert_eq!(t1.is_f64(), true);

let t2 = Type::f32();
assert_eq!(t2.is_f64(), false);

pub const fn is_array(&self) -> bool

Checks if the type is a Array.

let t1 = Type::bool();
assert_eq!(t1.is_array(), false);

pub const fn is_struct(&self) -> bool

Checks if the type is a Struct.

let t1 = Type::bool();
assert_eq!(t1.is_struct(), false);

pub const fn is_aggregate(&self) -> bool

Checks if the type is a Struct or an Array.

let t1 = Type::bool();
assert_eq!(t1.is_aggregate(), false);

pub fn as_bool(&self) -> Option<Bool>

If self models a boolean type, unwraps the boolean type and returns it.

pub fn as_ptr(&self) -> Option<Ptr>

If self models a pointer type, unwraps the pointer type and returns it.

pub fn as_int(&self) -> Option<Int>

If self models a integer type, unwraps the integer type and returns it.

pub fn as_float(&self) -> Option<Float>

If self models a float type, unwraps the float type and returns it.

pub fn as_array(&self) -> Option<Array>

If self models an array type, unwraps the array type and returns it.

pub fn as_struct(&self) -> Option<Struct>

If self models a struct type, unwraps the struct type and returns it.

pub fn unpack(self) -> UType

Unpacks a Type into an UType. This makes it take up more storage, but it also means that you can use native match and other pattern matching syntax on the type.

let ty = Type::bool();

match ty.unpack() {
    UType::Bool(_) => println!("got a bool!"),
    _ => {}
}

pub fn unwrap_bool(self) -> Bool

Equivalent to Self::as_bool, but panics on failure.

pub fn unwrap_ptr(self) -> Ptr

Equivalent to Self::as_ptr, but panics on failure.

pub fn unwrap_int(self) -> Int

Equivalent to Self::as_int, but panics on failure.

pub fn unwrap_float(self) -> Float

Equivalent to Self::as_float, but panics on failure.

pub fn unwrap_array(self) -> Array

Equivalent to Self::as_array, but panics on failure.

pub fn unwrap_struct(self) -> Struct

Equivalent to Self::as_struct, but panics on failure.