Trait sapphire::ir::BinaryInst

pub trait BinaryInst: Instruction {
    fn is_commutative(&self) -> bool;

    fn lhs(&self) -> Value { ... }
    fn rhs(&self) -> Value { ... }
}

Some instructions model binary operations, those instructions model this trait.

A valid assumption for any type implementing this trait is that the operands returned by Instruction::operands has a length of exactly 2. However, it is not valid to assume that any instruction implementing this follows the same pattern as most arithmetic instructions (i.e. Self::lhs and Self::rhs do not always have the same type, and Instruction::result_ty may be different from the type of the operands).

Required Methods

fn is_commutative(&self) -> bool

Checks if the binary instruction follows the commutative property, i.e. whether it is safe to swap the operands while preserving the behavior.

Provided Methods

fn lhs(&self) -> Value

Gets the left-hand operand of the instruction. For inst T %a, %b, this would be %a.

fn rhs(&self) -> Value

Gets the left-hand operand of the instruction. For inst T %a, %b, this would be %b.