Trait sapphire::ir::Cursor

pub trait Cursor: Sized {

    fn pos(&self) -> CursorPos;
    fn set_pos(&mut self, pos: CursorPos);
    fn def(&self) -> &FunctionDefinition;

    fn layout(&self) -> &Layout { ... }
    fn dfg(&self) -> &DataFlowGraph { ... }
    fn current_block(&self) -> Option<Block> { ... }
    fn current_inst(&self) -> Option<Inst> { ... }
    fn goto_before(&mut self, block: Block) { ... }
    fn goto_after(&mut self, block: Block) { ... }
    fn goto_first_inst(&mut self, block: Block) { ... }
    fn goto_last_inst(&mut self, block: Block) { ... }
    fn goto_inst(&mut self, inst: Inst) { ... }
    fn next_block(&mut self) -> Option<Block> { ... }
    fn prev_block(&mut self) -> Option<Block> { ... }
    fn next_inst(&mut self) -> Option<Inst> { ... }
    fn prev_inst(&mut self) -> Option<Inst> { ... }

}

Models basic cursor operations that view a function. None of these operations require mutable access to a given function, so they can be used inside of analyses.

Required Methods

fn pos(&self) -> CursorPos

Gets the current cursor position

fn set_pos(&mut self, pos: CursorPos)

Sets the current cursor position

fn def(&self) -> &FunctionDefinition

Returns the definition of the function being viewed

Provided Methods

fn layout(&self) -> &Layout

Gets the layout associated with the function being viewed

fn dfg(&self) -> &DataFlowGraph

Gets the data-flow graph associated with the function being viewed

fn current_block(&self) -> Option<Block>

Gets the current block being viewed by the cursor, if any.

fn current_inst(&self) -> Option<Inst>

Gets the current block being viewed by the cursor, if any.

fn goto_before(&mut self, block: Block)

Moves the position to Before(block).

fn goto_after(&mut self, block: Block)

Moves the position to After(block).

fn goto_first_inst(&mut self, block: Block)

Moves the position to At(block, first_inst_in_block).

fn goto_last_inst(&mut self, block: Block)

Moves the position to At(block, last_inst_in_block).

fn goto_inst(&mut self, inst: Inst)

Moves the position to At(containing, inst)

fn next_block(&mut self) -> Option<Block>

Moves the cursor to the next block in the function. If the cursor is currently not pointing to anywhere in the function, this moves it to Before(entry). If the cursor is pointing at the last block in the function, this moves it to Nothing.

fn prev_block(&mut self) -> Option<Block>

Moves the cursor to the block before the current one. If the cursor is pointing at Nothing, nothing changes. If the cursor is pointing at the first block, its moved to Nothing.

fn next_inst(&mut self) -> Option<Inst>

Moves the cursor to the next instruction in the function. If the cursor points before the block, this is the first instruction. If it points after, this does nothing. If it points at nothing, this does nothing.

fn prev_inst(&mut self) -> Option<Inst>

Moves the cursor to the previous instruction in the function. If the cursor points after the block, this is the last instruction. If it points before, this does nothing. If it points at nothing, this does nothing.