Trait sapphire::ir::CursorMut

pub trait CursorMut: Cursor {

    // Required methods
    fn def_mut(&mut self) -> &mut FunctionDefinition;
    fn func_mut(&mut self) -> &mut Function;

    // Provided methods
    fn replace(&mut self) -> ReplaceBuilder<'_> { ... }
    fn insert(&mut self) -> InsertBuilder<'_> { ... }
    fn replace_uses_with(&mut self, original: Value, new: Value) { ... }
    fn remove_inst(&mut self) -> Inst { ... }
    fn remove_inst_and_move_back(&mut self) -> Inst { ... }
    fn remove_block_param(&mut self, block: Block, param: Value) { ... }
    fn rewrite_branch_args(
        &mut self,
        branch: Inst,
        target: Block,
        new: &[Value]
    ) { ... }
    fn replace_branch_arg(
        &mut self,
        branch: Inst,
        target: Block,
        index: usize,
        new: Value
    ) { ... }
    fn remove_block(&mut self) { ... }
    fn remove_stack_slot(&mut self, slot: StackSlot) { ... }
    fn create_block(&mut self, name: &str) -> Block { ... }
    fn create_block_before(&mut self, name: &str, before: Block) -> Block { ... }
    fn create_block_after(&mut self, name: &str, after: Block) -> Block { ... }
    fn rewrite_branch_target(
        &mut self,
        to: Block,
        new: BlockWithParams
    ) -> BlockWithParams { ... }

}

A cursor with additional methods for mutating the IR.

Required Methods

fn def_mut(&mut self) -> &mut FunctionDefinition

Gets a mutable reference to the function’s definition.

fn func_mut(&mut self) -> &mut Function

Gets a mutable reference to the function itself

Provided Methods

fn replace(&mut self) -> ReplaceBuilder<'_>

Returns an instruction builder that replaces the current instruction.

fn insert(&mut self) -> InsertBuilder<'_>

Allows an instruction to be inserted before the current one being pointed at.

If this points at At(inst), an instruction is inserted before inst. If this is pointing at After(bb), an instruction is appended. Either way, the cursor is not moved, so repeated Self::insert calls make the instructions appear in the same order in the program.

fn replace_uses_with(&mut self, original: Value, new: Value)

Replaces any uses of original in the function with new.

fn remove_inst(&mut self) -> Inst

Removes the current instruction and leaves the cursor pointing at the instruction immediately following the one that was removed.

Returns the instruction that was removed.

fn remove_inst_and_move_back(&mut self) -> Inst

Removes the current instruction and leaves the cursor pointing at the instruction immediately before the one that was removed.

Returns the instruction that was removed.

fn remove_block_param(&mut self, block: Block, param: Value)

Removes a block parameter from a given block.

fn rewrite_branch_args(&mut self, branch: Inst, target: Block, new: &[Value])

Rewrites the arguments of a given branch to match new

fn replace_branch_arg( &mut self, branch: Inst, target: Block, index: usize, new: Value )

Rewrites a branch to target to have new as the indexth argument

fn remove_block(&mut self)

Removes the current block, and moves to the next one.

fn remove_stack_slot(&mut self, slot: StackSlot)

Removes a single stack slot.

fn create_block(&mut self, name: &str) -> Block

Creates a single basic block and returns it. This block is appended to the end of the block list.

Note that this does not switch the cursor to operate on that block, you still need to call goto_before or similar.

fn create_block_before(&mut self, name: &str, before: Block) -> Block

Equivalent to Self::create_block, except it inserts the block before before instead of appending it.

fn create_block_after(&mut self, name: &str, after: Block) -> Block

Equivalent to Self::create_block, except it inserts the block after after instead of appending it.

fn rewrite_branch_target( &mut self, to: Block, new: BlockWithParams ) -> BlockWithParams

Rewrites the branch at the end of the current block that targets to to instead branch to new (with the args associated with new).

Returns the old branch target so you can re-use the arguments.