Trait sapphire::ir::GenericInstVisitor

pub trait GenericInstVisitor<T> {

    fn visit_call(&mut self, data: &CallInst) -> T;
    fn visit_indirectcall(&mut self, data: &IndirectCallInst) -> T;
    fn visit_icmp(&mut self, data: &ICmpInst) -> T;
    fn visit_fcmp(&mut self, data: &FCmpInst) -> T;
    fn visit_sel(&mut self, data: &SelInst) -> T;
    fn visit_br(&mut self, data: &BrInst) -> T;
    fn visit_condbr(&mut self, data: &CondBrInst) -> T;
    fn visit_unreachable(&mut self, data: &UnreachableInst) -> T;
    fn visit_ret(&mut self, data: &RetInst) -> T;
    fn visit_and(&mut self, data: &CommutativeArithInst) -> T;
    fn visit_or(&mut self, data: &CommutativeArithInst) -> T;
    fn visit_xor(&mut self, data: &CommutativeArithInst) -> T;
    fn visit_shl(&mut self, data: &ArithInst) -> T;
    fn visit_ashr(&mut self, data: &ArithInst) -> T;
    fn visit_lshr(&mut self, data: &ArithInst) -> T;
    fn visit_iadd(&mut self, data: &CommutativeArithInst) -> T;
    fn visit_isub(&mut self, data: &ArithInst) -> T;
    fn visit_imul(&mut self, data: &CommutativeArithInst) -> T;
    fn visit_sdiv(&mut self, data: &ArithInst) -> T;
    fn visit_udiv(&mut self, data: &ArithInst) -> T;
    fn visit_srem(&mut self, data: &ArithInst) -> T;
    fn visit_urem(&mut self, data: &ArithInst) -> T;
    fn visit_fneg(&mut self, data: &FloatUnaryInst) -> T;
    fn visit_fadd(&mut self, data: &ArithInst) -> T;
    fn visit_fsub(&mut self, data: &ArithInst) -> T;
    fn visit_fmul(&mut self, data: &ArithInst) -> T;
    fn visit_fdiv(&mut self, data: &ArithInst) -> T;
    fn visit_frem(&mut self, data: &ArithInst) -> T;
    fn visit_alloca(&mut self, data: &AllocaInst) -> T;
    fn visit_load(&mut self, data: &LoadInst) -> T;
    fn visit_store(&mut self, data: &StoreInst) -> T;
    fn visit_offset(&mut self, data: &OffsetInst) -> T;
    fn visit_extract(&mut self, data: &ExtractInst) -> T;
    fn visit_insert(&mut self, data: &InsertInst) -> T;
    fn visit_elemptr(&mut self, data: &ElemPtrInst) -> T;
    fn visit_sext(&mut self, data: &CastInst) -> T;
    fn visit_zext(&mut self, data: &CastInst) -> T;
    fn visit_trunc(&mut self, data: &CastInst) -> T;
    fn visit_itob(&mut self, data: &CastInst) -> T;
    fn visit_btoi(&mut self, data: &CastInst) -> T;
    fn visit_sitof(&mut self, data: &CastInst) -> T;
    fn visit_uitof(&mut self, data: &CastInst) -> T;
    fn visit_ftosi(&mut self, data: &CastInst) -> T;
    fn visit_ftoui(&mut self, data: &CastInst) -> T;
    fn visit_fext(&mut self, data: &CastInst) -> T;
    fn visit_ftrunc(&mut self, data: &CastInst) -> T;
    fn visit_itop(&mut self, data: &CastInst) -> T;
    fn visit_ptoi(&mut self, data: &CastInst) -> T;
    fn visit_iconst(&mut self, data: &IConstInst) -> T;
    fn visit_fconst(&mut self, data: &FConstInst) -> T;
    fn visit_bconst(&mut self, data: &BConstInst) -> T;
    fn visit_undef(&mut self, data: &UndefConstInst) -> T;
    fn visit_null(&mut self, data: &NullConstInst) -> T;
    fn visit_globaladdr(&mut self, data: &GlobalAddrInst) -> T;

    fn dispatch_inst(&mut self, data: &InstData) -> T { ... }

}

A trait for a generic “visit an instruction” type. This is the smallest-scale visitor, as this visitor isn’t even aware of the function that a given instruction is in.

Required Methods

fn visit_call(&mut self, data: &CallInst) -> T

Visits a call instruction.

fn visit_indirectcall(&mut self, data: &IndirectCallInst) -> T

Visits an indirectcall instruction.

fn visit_icmp(&mut self, data: &ICmpInst) -> T

Visits an icmp instruction.

fn visit_fcmp(&mut self, data: &FCmpInst) -> T

Visits an fcmp instruction.

fn visit_sel(&mut self, data: &SelInst) -> T

Visits a sel instruction.

fn visit_br(&mut self, data: &BrInst) -> T

Visits a br instruction.

fn visit_condbr(&mut self, data: &CondBrInst) -> T

Visits a condbr instruction.

fn visit_unreachable(&mut self, data: &UnreachableInst) -> T

Visits an unreachable instruction.

fn visit_ret(&mut self, data: &RetInst) -> T

Visits a ret instruction.

fn visit_and(&mut self, data: &CommutativeArithInst) -> T

Visits an and instruction.

fn visit_or(&mut self, data: &CommutativeArithInst) -> T

Visits an or instruction.

fn visit_xor(&mut self, data: &CommutativeArithInst) -> T

Visits an xor instruction.

fn visit_shl(&mut self, data: &ArithInst) -> T

Visits a shl instruction.

fn visit_ashr(&mut self, data: &ArithInst) -> T

Visits an ashr instruction.

fn visit_lshr(&mut self, data: &ArithInst) -> T

Visits a lshr instruction.

fn visit_iadd(&mut self, data: &CommutativeArithInst) -> T

Visits an iadd instruction.

fn visit_isub(&mut self, data: &ArithInst) -> T

Visits an isub instruction.

fn visit_imul(&mut self, data: &CommutativeArithInst) -> T

Visits an imul instruction.

fn visit_sdiv(&mut self, data: &ArithInst) -> T

Visits an sdiv instruction.

fn visit_udiv(&mut self, data: &ArithInst) -> T

Visits an udiv instruction.

fn visit_srem(&mut self, data: &ArithInst) -> T

Visits a srem instruction.

fn visit_urem(&mut self, data: &ArithInst) -> T

Visits a urem instruction.

fn visit_fneg(&mut self, data: &FloatUnaryInst) -> T

Visits an fneg instruction.

fn visit_fadd(&mut self, data: &ArithInst) -> T

Visits an fadd instruction.

fn visit_fsub(&mut self, data: &ArithInst) -> T

Visits an fsub instruction.

fn visit_fmul(&mut self, data: &ArithInst) -> T

Visits an fmul instruction.

fn visit_fdiv(&mut self, data: &ArithInst) -> T

Visits an fdiv instruction.

fn visit_frem(&mut self, data: &ArithInst) -> T

Visits an frem instruction.

fn visit_alloca(&mut self, data: &AllocaInst) -> T

Visits an alloca instruction.

fn visit_load(&mut self, data: &LoadInst) -> T

Visits a load instruction.

fn visit_store(&mut self, data: &StoreInst) -> T

Visits a store instruction.

fn visit_offset(&mut self, data: &OffsetInst) -> T

Visits an offset instruction.

fn visit_extract(&mut self, data: &ExtractInst) -> T

Visits an extract instruction.

fn visit_insert(&mut self, data: &InsertInst) -> T

Visits an insert instruction.

fn visit_elemptr(&mut self, data: &ElemPtrInst) -> T

Visits an elemptr instruction.

fn visit_sext(&mut self, data: &CastInst) -> T

Visits a sext instruction.

fn visit_zext(&mut self, data: &CastInst) -> T

Visits a zext instruction.

fn visit_trunc(&mut self, data: &CastInst) -> T

Visits a trunc instruction.

fn visit_itob(&mut self, data: &CastInst) -> T

Visits an itob instruction.

fn visit_btoi(&mut self, data: &CastInst) -> T

Visits a btoi instruction.

fn visit_sitof(&mut self, data: &CastInst) -> T

Visits a sitof instruction.

fn visit_uitof(&mut self, data: &CastInst) -> T

Visits a uitof instruction.

fn visit_ftosi(&mut self, data: &CastInst) -> T

Visits an ftosi instruction.

fn visit_ftoui(&mut self, data: &CastInst) -> T

Visits an ftoui instruction.

fn visit_fext(&mut self, data: &CastInst) -> T

Visits an fext instruction.

fn visit_ftrunc(&mut self, data: &CastInst) -> T

Visits an ftrunc instruction.

fn visit_itop(&mut self, data: &CastInst) -> T

Visits an itop instruction.

fn visit_ptoi(&mut self, data: &CastInst) -> T

Visits a ptoi instruction.

fn visit_iconst(&mut self, data: &IConstInst) -> T

Visits an iconst instruction.

fn visit_fconst(&mut self, data: &FConstInst) -> T

Visits an fconst instruction.

fn visit_bconst(&mut self, data: &BConstInst) -> T

Visits a bconst instruction.

fn visit_undef(&mut self, data: &UndefConstInst) -> T

Visits an undef instruction.

fn visit_null(&mut self, data: &NullConstInst) -> T

Visits a null instruction.

fn visit_globaladdr(&mut self, data: &GlobalAddrInst) -> T

Visits a globaladdr instruction.

Provided Methods

fn dispatch_inst(&mut self, data: &InstData) -> T

Dispatcher that does the default behavior of calling the most specific visitor for each instruction.