Trait sapphire::ir::SIRVisitor

pub trait SIRVisitor<'a> {

    fn module(&self) -> &'a Module;
    fn visit_call(&mut self, inst: Inst, data: &CallInst, def: &FunctionDefinition);
    fn visit_indirectcall(
        &mut self,
        inst: Inst,
        data: &IndirectCallInst,
        def: &FunctionDefinition
    );
    fn visit_icmp(&mut self, inst: Inst, data: &ICmpInst, def: &FunctionDefinition);
    fn visit_fcmp(&mut self, inst: Inst, data: &FCmpInst, def: &FunctionDefinition);
    fn visit_sel(&mut self, inst: Inst, data: &SelInst, def: &FunctionDefinition);
    fn visit_br(&mut self, inst: Inst, data: &BrInst, def: &FunctionDefinition);
    fn visit_condbr(
        &mut self,
        inst: Inst,
        data: &CondBrInst,
        def: &FunctionDefinition
    );
    fn visit_unreachable(
        &mut self,
        inst: Inst,
        data: &UnreachableInst,
        def: &FunctionDefinition
    );
    fn visit_ret(&mut self, inst: Inst, data: &RetInst, def: &FunctionDefinition);
    fn visit_and(
        &mut self,
        inst: Inst,
        data: &CommutativeArithInst,
        def: &FunctionDefinition
    );
    fn visit_or(
        &mut self,
        inst: Inst,
        data: &CommutativeArithInst,
        def: &FunctionDefinition
    );
    fn visit_xor(
        &mut self,
        inst: Inst,
        data: &CommutativeArithInst,
        def: &FunctionDefinition
    );
    fn visit_shl(&mut self, inst: Inst, data: &ArithInst, def: &FunctionDefinition);
    fn visit_ashr(
        &mut self,
        inst: Inst,
        data: &ArithInst,
        def: &FunctionDefinition
    );
    fn visit_lshr(
        &mut self,
        inst: Inst,
        data: &ArithInst,
        def: &FunctionDefinition
    );
    fn visit_iadd(
        &mut self,
        inst: Inst,
        data: &CommutativeArithInst,
        def: &FunctionDefinition
    );
    fn visit_isub(
        &mut self,
        inst: Inst,
        data: &ArithInst,
        def: &FunctionDefinition
    );
    fn visit_imul(
        &mut self,
        inst: Inst,
        data: &CommutativeArithInst,
        def: &FunctionDefinition
    );
    fn visit_sdiv(
        &mut self,
        inst: Inst,
        data: &ArithInst,
        def: &FunctionDefinition
    );
    fn visit_udiv(
        &mut self,
        inst: Inst,
        data: &ArithInst,
        def: &FunctionDefinition
    );
    fn visit_srem(
        &mut self,
        inst: Inst,
        data: &ArithInst,
        def: &FunctionDefinition
    );
    fn visit_urem(
        &mut self,
        inst: Inst,
        data: &ArithInst,
        def: &FunctionDefinition
    );
    fn visit_fneg(
        &mut self,
        inst: Inst,
        data: &FloatUnaryInst,
        def: &FunctionDefinition
    );
    fn visit_fadd(
        &mut self,
        inst: Inst,
        data: &ArithInst,
        def: &FunctionDefinition
    );
    fn visit_fsub(
        &mut self,
        inst: Inst,
        data: &ArithInst,
        def: &FunctionDefinition
    );
    fn visit_fmul(
        &mut self,
        inst: Inst,
        data: &ArithInst,
        def: &FunctionDefinition
    );
    fn visit_fdiv(
        &mut self,
        inst: Inst,
        data: &ArithInst,
        def: &FunctionDefinition
    );
    fn visit_frem(
        &mut self,
        inst: Inst,
        data: &ArithInst,
        def: &FunctionDefinition
    );
    fn visit_alloca(
        &mut self,
        inst: Inst,
        data: &AllocaInst,
        def: &FunctionDefinition
    );
    fn visit_load(&mut self, inst: Inst, data: &LoadInst, def: &FunctionDefinition);
    fn visit_store(
        &mut self,
        inst: Inst,
        data: &StoreInst,
        def: &FunctionDefinition
    );
    fn visit_offset(
        &mut self,
        inst: Inst,
        data: &OffsetInst,
        def: &FunctionDefinition
    );
    fn visit_extract(
        &mut self,
        inst: Inst,
        data: &ExtractInst,
        def: &FunctionDefinition
    );
    fn visit_insert(
        &mut self,
        inst: Inst,
        data: &InsertInst,
        def: &FunctionDefinition
    );
    fn visit_elemptr(
        &mut self,
        inst: Inst,
        data: &ElemPtrInst,
        def: &FunctionDefinition
    );
    fn visit_sext(&mut self, inst: Inst, data: &CastInst, def: &FunctionDefinition);
    fn visit_zext(&mut self, inst: Inst, data: &CastInst, def: &FunctionDefinition);
    fn visit_trunc(
        &mut self,
        inst: Inst,
        data: &CastInst,
        def: &FunctionDefinition
    );
    fn visit_itob(&mut self, inst: Inst, data: &CastInst, def: &FunctionDefinition);
    fn visit_btoi(&mut self, inst: Inst, data: &CastInst, def: &FunctionDefinition);
    fn visit_sitof(
        &mut self,
        inst: Inst,
        data: &CastInst,
        def: &FunctionDefinition
    );
    fn visit_uitof(
        &mut self,
        inst: Inst,
        data: &CastInst,
        def: &FunctionDefinition
    );
    fn visit_ftosi(
        &mut self,
        inst: Inst,
        data: &CastInst,
        def: &FunctionDefinition
    );
    fn visit_ftoui(
        &mut self,
        inst: Inst,
        data: &CastInst,
        def: &FunctionDefinition
    );
    fn visit_fext(&mut self, inst: Inst, data: &CastInst, def: &FunctionDefinition);
    fn visit_ftrunc(
        &mut self,
        inst: Inst,
        data: &CastInst,
        def: &FunctionDefinition
    );
    fn visit_itop(&mut self, inst: Inst, data: &CastInst, def: &FunctionDefinition);
    fn visit_ptoi(&mut self, inst: Inst, data: &CastInst, def: &FunctionDefinition);
    fn visit_iconst(
        &mut self,
        inst: Inst,
        data: &IConstInst,
        def: &FunctionDefinition
    );
    fn visit_fconst(
        &mut self,
        inst: Inst,
        data: &FConstInst,
        def: &FunctionDefinition
    );
    fn visit_bconst(
        &mut self,
        inst: Inst,
        data: &BConstInst,
        def: &FunctionDefinition
    );
    fn visit_undef(
        &mut self,
        inst: Inst,
        data: &UndefConstInst,
        def: &FunctionDefinition
    );
    fn visit_null(
        &mut self,
        inst: Inst,
        data: &NullConstInst,
        def: &FunctionDefinition
    );
    fn visit_globaladdr(
        &mut self,
        inst: Inst,
        data: &GlobalAddrInst,
        def: &FunctionDefinition
    );

    fn dispatch_funcs(&mut self) { ... }
    fn walk(&mut self) { ... }
    fn dispatch_blocks(&mut self, def: &FunctionDefinition) { ... }
    fn visit_func(&mut self, func: Func) { ... }
    fn dispatch_insts(&mut self, block: Block, def: &FunctionDefinition) { ... }
    fn visit_block(&mut self, block: Block, def: &FunctionDefinition) { ... }
    fn dispatch_inst(
        &mut self,
        inst: Inst,
        data: &InstData,
        def: &FunctionDefinition
    ) { ... }
    fn visit_inst(&mut self, inst: Inst, def: &FunctionDefinition) { ... }

}

Trait that allows configurable and simple SIR visiting.

This trait is for use when basically anything in the IR needs to be matched somehow, when there are a few very specific patterns that need to be matched this is not the right API.

Required Methods

fn module(&self) -> &'a Module

Gets the module being visited.

fn visit_call(&mut self, inst: Inst, data: &CallInst, def: &FunctionDefinition)

Visits a call instruction.

fn visit_indirectcall(
&mut self,
inst: Inst,
data: &IndirectCallInst,
def: &FunctionDefinition
)

Visits an indirectcall instruction.

fn visit_icmp(&mut self, inst: Inst, data: &ICmpInst, def: &FunctionDefinition)

Visits an icmp instruction.

fn visit_fcmp(&mut self, inst: Inst, data: &FCmpInst, def: &FunctionDefinition)

Visits an fcmp instruction.

fn visit_sel(&mut self, inst: Inst, data: &SelInst, def: &FunctionDefinition)

Visits a sel instruction.

fn visit_br(&mut self, inst: Inst, data: &BrInst, def: &FunctionDefinition)

Visits a br instruction.

fn visit_condbr(&mut self, inst: Inst, data: &CondBrInst, def: &FunctionDefinition)

Visits a condbr instruction.

fn visit_unreachable(
&mut self,
inst: Inst,
data: &UnreachableInst,
def: &FunctionDefinition
)

Visits an unreachable instruction.

fn visit_ret(&mut self, inst: Inst, data: &RetInst, def: &FunctionDefinition)

Visits a ret instruction.

fn visit_and(
&mut self,
inst: Inst,
data: &CommutativeArithInst,
def: &FunctionDefinition
)

Visits an and instruction.

fn visit_or(
&mut self,
inst: Inst,
data: &CommutativeArithInst,
def: &FunctionDefinition
)

Visits an or instruction.

fn visit_xor(
&mut self,
inst: Inst,
data: &CommutativeArithInst,
def: &FunctionDefinition
)

Visits an xor instruction.

fn visit_shl(&mut self, inst: Inst, data: &ArithInst, def: &FunctionDefinition)

Visits a shl instruction.

fn visit_ashr(&mut self, inst: Inst, data: &ArithInst, def: &FunctionDefinition)

Visits an ashr instruction.

fn visit_lshr(&mut self, inst: Inst, data: &ArithInst, def: &FunctionDefinition)

Visits a lshr instruction.

fn visit_iadd(
&mut self,
inst: Inst,
data: &CommutativeArithInst,
def: &FunctionDefinition
)

Visits an iadd instruction.

fn visit_isub(&mut self, inst: Inst, data: &ArithInst, def: &FunctionDefinition)

Visits an isub instruction.

fn visit_imul(
&mut self,
inst: Inst,
data: &CommutativeArithInst,
def: &FunctionDefinition
)

Visits an imul instruction.

fn visit_sdiv(&mut self, inst: Inst, data: &ArithInst, def: &FunctionDefinition)

Visits an sdiv instruction.

fn visit_udiv(&mut self, inst: Inst, data: &ArithInst, def: &FunctionDefinition)

Visits an udiv instruction.

fn visit_srem(&mut self, inst: Inst, data: &ArithInst, def: &FunctionDefinition)

Visits a srem instruction.

fn visit_urem(&mut self, inst: Inst, data: &ArithInst, def: &FunctionDefinition)

Visits a urem instruction.

fn visit_fneg(
&mut self,
inst: Inst,
data: &FloatUnaryInst,
def: &FunctionDefinition
)

Visits an fneg instruction.

fn visit_fadd(&mut self, inst: Inst, data: &ArithInst, def: &FunctionDefinition)

Visits an fadd instruction.

fn visit_fsub(&mut self, inst: Inst, data: &ArithInst, def: &FunctionDefinition)

Visits an fsub instruction.

fn visit_fmul(&mut self, inst: Inst, data: &ArithInst, def: &FunctionDefinition)

Visits an fmul instruction.

fn visit_fdiv(&mut self, inst: Inst, data: &ArithInst, def: &FunctionDefinition)

Visits an fdiv instruction.

fn visit_frem(&mut self, inst: Inst, data: &ArithInst, def: &FunctionDefinition)

Visits an frem instruction.

fn visit_alloca(&mut self, inst: Inst, data: &AllocaInst, def: &FunctionDefinition)

Visits an alloca instruction.

fn visit_load(&mut self, inst: Inst, data: &LoadInst, def: &FunctionDefinition)

Visits a load instruction.

fn visit_store(&mut self, inst: Inst, data: &StoreInst, def: &FunctionDefinition)

Visits a store instruction.

fn visit_offset(&mut self, inst: Inst, data: &OffsetInst, def: &FunctionDefinition)

Visits an offset instruction.

fn visit_extract(
&mut self,
inst: Inst,
data: &ExtractInst,
def: &FunctionDefinition
)

Visits an extract instruction.

fn visit_insert(&mut self, inst: Inst, data: &InsertInst, def: &FunctionDefinition)

Visits an insert instruction.

fn visit_elemptr(
&mut self,
inst: Inst,
data: &ElemPtrInst,
def: &FunctionDefinition
)

Visits an elemptr instruction.

fn visit_sext(&mut self, inst: Inst, data: &CastInst, def: &FunctionDefinition)

Visits a sext instruction.

fn visit_zext(&mut self, inst: Inst, data: &CastInst, def: &FunctionDefinition)

Visits a zext instruction.

fn visit_trunc(&mut self, inst: Inst, data: &CastInst, def: &FunctionDefinition)

Visits a trunc instruction.

fn visit_itob(&mut self, inst: Inst, data: &CastInst, def: &FunctionDefinition)

Visits an itob instruction.

fn visit_btoi(&mut self, inst: Inst, data: &CastInst, def: &FunctionDefinition)

Visits a btoi instruction.

fn visit_sitof(&mut self, inst: Inst, data: &CastInst, def: &FunctionDefinition)

Visits a sitof instruction.

fn visit_uitof(&mut self, inst: Inst, data: &CastInst, def: &FunctionDefinition)

Visits a uitof instruction.

fn visit_ftosi(&mut self, inst: Inst, data: &CastInst, def: &FunctionDefinition)

Visits an ftosi instruction.

fn visit_ftoui(&mut self, inst: Inst, data: &CastInst, def: &FunctionDefinition)

Visits an ftoui instruction.

fn visit_fext(&mut self, inst: Inst, data: &CastInst, def: &FunctionDefinition)

Visits an fext instruction.

fn visit_ftrunc(&mut self, inst: Inst, data: &CastInst, def: &FunctionDefinition)

Visits an ftrunc instruction.

fn visit_itop(&mut self, inst: Inst, data: &CastInst, def: &FunctionDefinition)

Visits an itop instruction.

fn visit_ptoi(&mut self, inst: Inst, data: &CastInst, def: &FunctionDefinition)

Visits a ptoi instruction.

fn visit_iconst(&mut self, inst: Inst, data: &IConstInst, def: &FunctionDefinition)

Visits an iconst instruction.

fn visit_fconst(&mut self, inst: Inst, data: &FConstInst, def: &FunctionDefinition)

Visits an fconst instruction.

fn visit_bconst(&mut self, inst: Inst, data: &BConstInst, def: &FunctionDefinition)

Visits a bconst instruction.

fn visit_undef(
&mut self,
inst: Inst,
data: &UndefConstInst,
def: &FunctionDefinition
)

Visits an undef instruction.

fn visit_null(
&mut self,
inst: Inst,
data: &NullConstInst,
def: &FunctionDefinition
)

Visits a null instruction.

fn visit_globaladdr(
&mut self,
inst: Inst,
data: &GlobalAddrInst,
def: &FunctionDefinition
)

Visits a globaladdr instruction.

Provided Methods

fn dispatch_funcs(&mut self)

Dispatcher that does the default walking behavior to every function in the module

fn walk(&mut self)

Walks over the module and calls the expected visit_* methods

fn dispatch_blocks(&mut self, def: &FunctionDefinition)

Dispatcher that does the default walking behavior, going to every block in program order.

fn visit_func(&mut self, func: Func)

Called whenever an individual function is visited.

fn dispatch_insts(&mut self, block: Block, def: &FunctionDefinition)

Dispatcher that does the default behavior of visiting every instruction in a given block in program order.

fn visit_block(&mut self, block: Block, def: &FunctionDefinition)

Called whenever an individual block is visited.

fn dispatch_inst(&mut self, inst: Inst, data: &InstData, def: &FunctionDefinition)

Dispatcher that does the default behavior of calling the most specific visitor for each instruction.

fn visit_inst(&mut self, inst: Inst, def: &FunctionDefinition)

Called whenever an individual instruction is visited.