Trait sapphire::ir::Terminator

pub trait Terminator: Instruction {
    // Required method
    fn targets(&self) -> &[BlockWithParams];
}

Models a terminator, i.e. the only instructions that are allowed at the end of a basic block.

All terminators transfer control flow somewhere unless they end execution, so users need to be able to query where control could be transferred to.

Required Methods

fn targets(&self) -> &[BlockWithParams]

Gets the possible blocks where control could be transferred to once this instruction is executed.

Note that this might be empty, see unreachable or ret.