Trait sapphire::ir::UnaryInst

pub trait UnaryInst: Instruction {
    // Provided method
    fn operand(&self) -> Value { ... }
}

Some instructions model unary operations, those instructions model this trait.

A valid assumption for any type implementing this trait is that the operands returned by Instruction::operands has a length of exactly 1.

While most instructions implementing this are casts, not all of them are. A few instructions (notably fneg) are unary but not casts.

Provided Methods

fn operand(&self) -> Value

Gets the single unary operand of the instruction. For inst T %a, this would be %a.