Macro sapphire::arena_key

macro_rules! arena_key {
    ( $(#[$outer:meta])* $vis:vis struct $name:ident($ty:ty); $($rest:tt)* ) => { ... };
    ( $(#[$outer:meta])* $vis:vis struct $name:ident; $($rest:tt)* ) => { ... };
    () => { ... };
}

Creates a type-safe key for a ArenaMap and associated data structures.

This macro allows the inner storage type of the key to be customized, or a default that should always work can be used if the inner type is not specified.

arena_key! {
    /// We can have doc comments! This one uses the default data type.
    pub struct EntityRef;
   
    // this one is private, and uses u8 as the key type.
    struct ExtremelyDenseRef(u8);
}

type EntityMap<V> = ArenaMap<EntityRef, V>;
type ExtremelyDenseMap<V> = ArenaMap<ExtremelyDenseRef, V>;