Struct sapphire::pass::ModuleAnalysisManager

#[repr(transparent)]
pub struct ModuleAnalysisManager(_);

A lazy analysis manager for a module of SIR.

Analysis passes are registered through Self::add_analysis, and then can be later requested through Self::get. These are lazily recomputed as they are invalidated and requested through different passes.

When a pass wants the result of an analysis, it uses Self::get which will either return the computed result, or if the analysis is “invalid” it will compute the result, cache it, and then return it.

When a pass completes, it returns a list of preserved analyses (PreservedAnalyses), those are then passed to this, which marks the analyses that are not preserved as invalid.

Implementations

impl ModuleAnalysisManager

pub fn new() -> Self

Creates a new ModuleAnalysisManager.

This manager has no analyses registered, they need to be added with Self::add_analysis before they can be used by transform passes.

pub fn invalidate(&mut self, ir: &Module, preserved: &PreservedAnalyses)

Invalidates a set of analyses for a given module.

Any analysis not explicitly marked to be preserved in preserved is considered to be invalidated.

pub fn add_analysis<T: ModuleAnalysisPass>(&mut self, pass: T)

Registers a module analysis pass. The pass is not run, but the state for it to be used later is set up inside the manager.

You cannot use T in any of the other methods in this type without having called this one with the same T first, or else you’ll get a panic.

pub fn get<T: ModuleAnalysisPass>(&self, ir: &Module) -> Ref<'_, T::Result>

Lazily gets the result of an analysis. If the analysis has been invalidated, the result is re-computed, cached, and then returned.

pub fn initialize_for_ir(&mut self, ir: &Module)

Initializes the analysis manager for a given bit of IR.

This must be called at least once before Self::get is called for a given unit of IR, or the manager will panic.