Trait sapphire::passes::FunctionAnalysisPass

pub trait FunctionAnalysisPass: Any {
    type Result;

    fn expects_preserved(&self) -> &'static [TypeId];
    fn run(
        &mut self,
        func: &Function,
        am: &FunctionAnalysisManager
    ) -> Self::Result;
}

An analysis that runs over an entire function.

Required Associated Types

type Result

The result type of a given analysis.

Required Methods

fn expects_preserved(&self) -> &'static [TypeId]

Allows analyses to declare dependence on other analyses.

This directly implies that if T is preserved, all analyses referred to by T::expects_preserved are also assumed to be preserved.

This is checked in debug mode.

fn run(&mut self, func: &Function, am: &FunctionAnalysisManager) -> Self::Result

Performs the analysis and returns a computed result. This should not be an impure operation, running the analysis twice on the same input should produce the same result.