Trait sapphire::passes::ModuleAnalysisPass

pub trait ModuleAnalysisPass: Any {
    type Result;

    fn expects_preserved(&self) -> &'static [TypeId];
    fn run(&mut self, module: &Module, am: &ModuleAnalysisManager) -> Self::Result;
}

An analysis pass that runs over an entire module.

Required Associated Types

type Result

The result type of a given analysis.

Required Methods

fn expects_preserved(&self) -> &'static [TypeId]

Allows analyses to declare dependence on other analyses.

This directly implies that if T is preserved, all analyses referred to by T::expects_preserved are also assumed to be preserved.

This is checked in debug mode.

fn run(&mut self, module: &Module, am: &ModuleAnalysisManager) -> Self::Result

Performs the analysis and returns a computed result. This should not be an impure operation, running the analysis twice on the same input should produce the same result.