Trait sapphire::passes::ModuleTransformPass

pub trait ModuleTransformPass {
    fn run(
        &mut self,
        module: &mut Module,
        am: &ModuleAnalysisManager
    ) -> PreservedAnalyses;
}

Models a pass that possibly performs a transformation over an entire SIR module.

While the pass may not actually modify the IR, it has the ability to, and needs to declare what it changed (if anything) through PreservedAnalyses.

Required Methods

fn run(
&mut self,
module: &mut Module,
am: &ModuleAnalysisManager
) -> PreservedAnalyses

Performs the transformation over a given SIR module.

This function is expected to act as-if it was pure, i.e. calling the same pass multiple times on the same IR should produce equivalent IR each time and should return the same preserved analyses each time.