Module sapphire::reader2

A hand-written parser and lexer meant to replace the Pest-based generated parser.

It should be significantly faster in every case than the Pest version.

Structs

Error

A parse error of some sort, this can be formatted by the caller

Lex

A lexer for SIR that lazily produces tokens.

Parser

A parser that parses a string containing Sapphire IR.

TokPair

A token yielded by the lexer, containing line/col information as well as raw token data.

Enums

CompareOpcode

The comparison performed by a given icmp or fcmp instruction. They are in the same lex type because there is overlap between them.

FloatLiteral

Different types of (unparsed) floating-point literals

IntLiteral

Different types of (unparsed) integer literals

Opcode

An instruction opcode, e.g. the iadd in iadd i32 %0, %1.

Token

A single lex token

Functions

format_parse_error

Formats a reader2::Parser error into a human-readable string

format_verifier_error

Formats errors emitted by the verifier, this is also used by reader2.

parse_sir

Parses SIR source code using the reader2 parser.

Type Aliases

ParseResult

A typedef for a Result that has Error as the Err type