Enum sapphire::reader2::Token

pub enum Token<'a> {

    LocalIdent(&'a str),
    StackIdent(&'a str),
    GlobalIdent(&'a str),
    LabelIdent(&'a str),
    Bool,
    Ptr,
    I8,
    I16,
    I32,
    I64,
    F32,
    F64,
    Void,
    Variadic,
    Fn,
    CurlyOpen,
    CurlyClose,
    ParenOpen,
    ParenClose,
    SquareOpen,
    SquareClose,
    Eq,
    Colon,
    X,
    Comma,
    ByVal,
    NoAlias,
    NonNull,
    CCC,
    SysV,
    Win64,
    Volatile,
    Opcode(Opcode),
    CmpOpcode(CompareOpcode),
    IntLitHex(&'a str),
    IntLitBinary(&'a str),
    IntLitOctal(&'a str),
    IntLitDecimal(&'a str),
    FloatLitStandard(&'a str),
    FloatLitScientific(&'a str),
    FloatLitRaw(&'a str),
    FloatLitNaN,
    BoolLit(bool),
    Stack,
    Unknown(&'a str),

}

A single lex token

Variants

LocalIdent(&'a str)

A local identifier, %ident

StackIdent(&'a str)

A stack identifier, $ident

GlobalIdent(&'a str)

A global identifier, @ident

LabelIdent(&'a str)

A label identifier, ident

Bool

bool

Ptr

ptr

I8

i8

I16

i16

I32

i32

I64

i64

F32

f32

F64

f64

Void

void

Variadic

...

Fn

fn

CurlyOpen

{

CurlyClose

}

ParenOpen

(

ParenClose

)

SquareOpen

[

SquareClose

]

Eq

=

Colon

:

X

x

Comma

,

ByVal

byval

NoAlias

noalias

NonNull

nonnull

CCC

ccc

SysV

sysv

Win64

win64

Volatile

volatile

Opcode(Opcode)

An instruction opcode

CmpOpcode(CompareOpcode)

An opcode for icmp or fcmp

IntLitHex(&'a str)

An integer literal A hex literal, e.g. 0xdeadbeef or 0xDEADBEEF

IntLitBinary(&'a str)

A binary literal, e.g. 0b1010

IntLitOctal(&'a str)

An octal literal, e.g. 0o715

IntLitDecimal(&'a str)

A decimal literal, e.g. -1 or 42

FloatLitStandard(&'a str)

A floating-point literal Standard float notation, e.g. 3.141592

FloatLitScientific(&'a str)

C-style scientific notation, e.g. 0.3e9, 1.0e-9

FloatLitRaw(&'a str)

A way of representing raw byte values for floating-point values, e.g. 0xfpFFFFFFFFFFFFFFFF.

FloatLitNaN

NaN

BoolLit(bool)

A bool literal

Stack

stack

Unknown(&'a str)

An unknown token, almost certainly an error

Implementations

impl<'a> Token<'a>

pub fn as_int_lit(self) -> Option<IntLiteral<'a>>

If self is one of the integer literals, returns it in a nicer-to-work-with form.

Unfortunately Token has to work like this to fit in 24 bytes, if it just held IntLiteral in the variants it would be 32.

pub fn as_float_lit(self) -> Option<FloatLiteral<'a>>

If self is one of the float literals, returns it in a nicer-to-work-with form.

Unfortunately Token has to work like this to fit in 24 bytes, if it just held FloatLiteral in the variants it would be 32.