Module sapphire::transforms

Defines the “transform” passes in SIR’s infrastructure.

These are the passes that can (potentially) modify SIR, and don’t actually logically yield a result.

Some of these “transforms” are not actually transformations (e.g. the verify pass is a “transform pass” even though it manipulates no IR), but most of them are. All of them logically yield no result except the IR that exists after they run.

Modules

common

Common utilities that are needed between different transformations.

Structs

AggressiveDCEPass

An aggressive dead code elimination pass.

DominatorTreeWriterPass

Wrapper pass that prints out a textual representation of a DominatorTree.

GVNPass

A simple GVN (global value numbering) pass.

Mem2RegPass

A memory to SSA promotion pass.

ModuleWriterPass

This is a pass that writes out a textual representation of a module to a given stream.

SCCPPass

A Sparse Conditional Constant Propagation pass.

SimplifyCFGPass

A control-flow graph simplification pass.

SimplifyInstPass

A peephole optimization and algebraic simplification pass.

SplitCriticalEdgesPass

Performs critical-edge splitting.

VerifyModulePass

An IR validity verification pass.

Functions

aggressive_instruction_dce

Scans over an entire function and removes dead instructions and φ nodes.

gvn

Runs a global value-numbering algorithm over func to remove redundant expressions.

mem2reg

Promotes function-local memory (that doesn’t escape) into φ nodes where possible.

split_crit_edges

Performs critical edge splitting over a function. Each edge that is split has a dummy node inserted that contains only an unconditional branch to the block that was originally being targeted.

verify_module

Verifies that an entire module is valid SIR.

verify_module_panic

This is verify_module, except that it writes out any errors and then panics on failure.