Module sapphire::transforms::common

Common utilities that are needed between different transformations.

Structs

InPlaceConstantFolder

A generic “fold this instruction if possible” utility. It perform transforms similar to the following:

Functions

could_cause_pointer_escape

Checks if a given instruction could possibly cause a pointer to escape.

has_side_effect

Checks whether a given instruction possibly has a side effect.

pointer_escapees

Looks at an instruction and returns any of its pointer operands that are considered to have “escaped” the function’s control.

rewrite_and_remove_block_param

Removes a parameter from a block and rewrites all branches to that block to not include that parameter.

rewrite_pad_branch_argument

Rewrites a branch target to replace an argument with another.