Struct sapphire::transforms::common::InPlaceConstantFolder

pub struct InPlaceConstantFolder<'f, 'c> { /* private fields */ }

A generic “fold this instruction if possible” utility. It perform transforms similar to the following:

%0 = iconst i32 42
%1 = iconst i32 6
%2 = iadd i32 %0, %1 ; -> `%2 = iconst i32 48`

Note that this will fold illegal operations (e.g. a load null or sdiv 1, 0) into unreachables.

Notable exceptions that it will never fold are floating-point operations due to the target-dependent behavior.

Note that this is an extremely naive transform, it will happily generate the same constant over and over again. A GVN and a DCE pass after using this will likely be necessary.

Implementations

impl<'f, 'c> InPlaceConstantFolder<'f, 'c>

pub fn try_fold( inst: Inst, data: &InstData, cursor: &'c mut FuncCursor<'f> ) -> bool

Attempts to perform trivial constant folding on a single instruction.

Returns whether or not the folder was able to make a simplification.