Struct sapphire::transforms::DeadCodeEliminationPass

pub struct DeadCodeEliminationPass;

Aggressive dead code elimination.

This algorithm iterates over a function in postorder, and sees all uses before definitions. It assumes everything (that can’t cause side effects) is dead until proven otherwise. It removes everything that it decides is dead.

This will also remove dead φ nodes from basic blocks, and will remove unreachable blocks entirely.