Module sapphire::utility

Provides several utility APIs that are used inside of various modules inside of the compiler.

This is the general catch-all for random utility code.

Structs

IntoIter

Implements IntoIter for a TinyArray.

PackedOption

Provides an Option-like type for Packable objects without any extra cost to store the flag. It takes up exactly as much space as T would on its own, relying on the [`Packable::reserved

RawSpinMutex

Effectively the same thing as SpinMutex, except it doesn’t call core::hint::spin_loop. This avoids potential tiny slowdowns from a pause or yield instruction, but also harms power efficiency and hyper-threading.

SpinMutex

A TTAS (test and test-and-set) spin-lock.

Str

A reference to a string inside of a given StringPool. These are significantly more compact than both Strings and &strs, and are thus better for usage inside IR storage where space is precious.

StringPool

Contains a number of heap-allocated strings, and provides an API to map Strs to those heap-allocated strings. All strings are automatically de-duplicated internally, so two Strs from two calls to Self::insert with the same string value will be equivalent.

TinyArray

A tiny, specialized container for one purpose: storing immutable arrays into data structures that need to be compact. This is effectively a small-size optimized [Box<[T]>].

Traits

IntoTree

Models a type that can be traversed in a tree-like fashion. This is intended for debug APIs / passes that deal in human-readable trees.

Packable

Helper trait for a type that can be packed into a PackedOption.

Functions

stringify_tree

Prints a tree in a consistent format.

Type Definitions

SaHashMap

Alias for std::collections::HashMap<K, V, ahash::RandomState>. This is a hash table that uses a faster hash function, this is significantly faster for the small keys that we are using everywhere.

SaHashSet

Alias for std::collections::HashSet<V, ahash::RandomState>. This is a hash table that uses a faster hash function, this is significantly faster for the small keys that we are using everywhere.