Struct sapphire::utility::StringPool

pub struct StringPool { /* private fields */ }

Contains a number of heap-allocated strings, and provides an API to map Strs to those heap-allocated strings. All strings are automatically de-duplicated internally, so two Strs from two calls to Self::insert with the same string value will be equivalent.

Note that a pool provides no protection against using Strs for the wrong pool, as for data compactness reasons that is simply not stored.

let mut pool = StringPool::new();

let k1 = pool.insert("Hello!");
let k2 = pool.insert("Goodbye!");

assert_eq!(&pool[k1], "Hello!");
assert_eq!(&pool[k2], "Goodbye!");

let k3 = pool.insert("Hello!");
assert_eq!(k1, k3);
assert_eq!(&pool[k1], "Hello!");

Implementations

impl StringPool

pub fn new() -> Self

Creates an empty string pool that has no valid indices into it.

let pool = StringPool::new();

pub fn insert(&mut self, string: &str) -> Str

Inserts a string into the pool and returns a reference that can be used to access the string.

let mut pool = StringPool::new();
let name = pool.insert("John Smith");
assert_eq!(&pool[name], "John Smith");

pub fn get(&self, index: Str) -> Option<&str>

Gets the string at a given Str index. If the index somehow doesn’t exist, None is returned.

let mut pool = StringPool::new();
let s1 = pool.insert("one");
assert_eq!(pool.get(s1), Some("one"));

pub fn values(&self) -> impl Iterator<Item = &str>

Provides an iterator over all of the strings that are in the pool.

let mut pool = StringPool::new();
let _ = pool.insert("Hello!");
let mut it = pool.values();
assert_eq!(it.next(), Some("Hello!"));

pub fn len(&self) -> usize

Returns the number of unique strings stored inside the pool.

let mut pool = StringPool::new();
pool.insert("a");
pool.insert("b");
pool.insert("b");
assert_eq!(pool.len(), 2);

pub fn is_empty(&self) -> bool

Checks if the pool contains any strings.

let mut pool = StringPool::new();
assert_eq!(pool.is_empty(), true);