Trait sapphire::utility::IntoTree

pub trait IntoTree<'a> {
    type Node: Copy;

    fn root(&'a self) -> Self::Node;
    fn children(&'a self, node: Self::Node) -> Vec<Self::Node>;
}

Models a type that can be traversed in a tree-like fashion. This is intended for debug APIs / passes that deal in human-readable trees.

This is not for anything that is actually in the hot path of any compiler passes, this is meant for human-facing APIs. What this means is that it may be extremely slow.

Required Associated Types

type Node: Copy

The node type of the tree

Required Methods

fn root(&'a self) -> Self::Node

Returns the root tree node

fn children(&'a self, node: Self::Node) -> Vec<Self::Node>

Returns the list of children that a given node has