Module sapphire::arena

A simple typed arena module.

These arenas mostly do not allow deletion (besides SecondaryMap) and provide configurable index sizes for extra flexibility and performance. It is used extensively for forming graphs and other complex data structures needed inside the compiler for representing different IRs.

Very similar to id_arena and other simple typed arena crates, except this one ties in better with the specific needs of this compiler (and does less safety checks in exchange for significantly reduced overhead per slot).

arena_key! {
    pub struct Node;
}

enum AstNode {
    Immediate(u64),
    Add(Node, Node),
    Mul(Node, Node)
}

let mut arena = ArenaMap::new();

// (16 + 3) * 3
let e1: Node = arena.insert(AstNode::Immediate(16)); // => 16
let e2 = arena.insert(AstNode::Immediate(3)); // => 3
let e3 = arena.insert(AstNode::Add(e1, e2)); // => (16 + 3)
let e4 = arena.insert(AstNode::Mul(e2, e3)); // => (16 + 3) * 3

Structs

ArenaMap

This is meant to act as a primary mapping of K -> V, where K is some key type and V is the value being stored. Other mappings that use the same key as an existing ArenaMap should use SecondaryMap instead.

IntoIter

Helper type for .into_iter() for arena map types

Keys

Provides a way of iterating over all of the keys in a given ArenaMap.

SecondaryMap

Intended to be a dense secondary mapping K -> V for keys from a primary ArenaMap. This is to associate extra data with most (but ideally all) keys from a given PrimaryMap.

UniqueArenaMap

Contains a table of immutable, unique elements. All elements are only stored once, no duplicates are stored inside of the arena itself. Lookups by key are almost as efficient as they are in a plain ArenaMap, although an extra pointer dereference must be done.

Traits

ArenaKey

Models a type that can act as a key for the arena map types.