Trait sapphire::arena::ArenaKey

pub trait ArenaKey: Copy + Eq + Debug {
    type Item;

    // Required methods
    fn key_new(index: usize) -> Self;
    fn key_index(self) -> usize;
}

Models a type that can act as a key for the arena map types.

This arena key can have a configurable size, and can lead to panics if the underlying data type is too small to represent a given index.

Note that in most cases this trait should not be implemented directly, prefer to use the arena_key or dense_arena_key macros that provide the implementation for you.

Required Associated Types

type Item

The underlying data type of the key.

Required Methods

fn key_new(index: usize) -> Self

Creates a new key from a given arena index. This should do any necessary conversions to convert a usize index into the internal storage type, ideally the conversion should be lossless.

This should check that the index is in the bounds of its internal storage type, but this is not required for safety. It will just be harder to track down bugs when indexes overflow and end up referring to the wrong key instead of panicking.

fn key_index(self) -> usize

Converts the internal storage type into a usize index.

This conversion should be lossless.