Trait sapphire::ir::Cursor

pub trait Cursor: Sized {

    // Required methods
    fn pos(&self) -> CursorPos;
    fn set_pos(&mut self, pos: CursorPos);
    fn def(&self) -> &FunctionDefinition;

    // Provided methods
    fn layout(&self) -> &Layout { ... }
    fn dfg(&self) -> &DataFlowGraph { ... }
    fn current_block(&self) -> Option<Block> { ... }
    fn current_inst(&self) -> Option<Inst> { ... }
    fn current_block_terminator_targets(&self) -> Option<&[BlockWithParams]> { ... }
    fn current_inst_dbg(&self) -> Option<DebugInfo> { ... }
    fn goto_function_begin(&mut self) { ... }
    fn goto_before(&mut self, block: Block) { ... }
    fn goto_after(&mut self, block: Block) { ... }
    fn goto_first_inst(&mut self, block: Block) { ... }
    fn goto_last_inst(&mut self, block: Block) { ... }
    fn goto_inst(&mut self, inst: Inst) { ... }
    fn next_block(&mut self) -> Option<Block> { ... }
    fn prev_block(&mut self) -> Option<Block> { ... }
    fn next_inst(&mut self) -> Option<Inst> { ... }
    fn prev_inst(&mut self) -> Option<Inst> { ... }
    fn inst_to_result(&self, inst: Inst) -> Option<Value> { ... }
    fn value_to_inst(&self, val: Value) -> Option<Inst> { ... }
    fn is_entry_block(&self, block: Block) -> bool { ... }
    fn entry_block(&self) -> Option<Block> { ... }
    fn ty(&self, value: Value) -> Type { ... }
    fn block_name(&self, block: Block) -> Str { ... }
    fn block_params(&self, block: Block) -> &[Value] { ... }
    fn value_def(&self, value: Value) -> ValueDef { ... }
    fn inst_data(&self, inst: Inst) -> &InstData { ... }
    fn inst_block(&self, inst: Inst) -> Block { ... }
    fn branch_info(&self, inst: Inst) -> Option<&[BlockWithParams]> { ... }
    fn inst_debug(&self, inst: Inst) -> DebugInfo { ... }
    fn val_debug(&self, val: Value) -> DebugInfo { ... }
    fn stack_slot(&self, slot: StackSlot) -> StackSlotData { ... }

}

Models basic cursor operations that view a function. None of these operations require mutable access to a given function, so they can be used inside of analyses.

Required Methods

fn pos(&self) -> CursorPos

Gets the current cursor position

fn set_pos(&mut self, pos: CursorPos)

Sets the current cursor position

fn def(&self) -> &FunctionDefinition

Returns the definition of the function being viewed

Provided Methods

fn layout(&self) -> &Layout

Gets the layout associated with the function being viewed

fn dfg(&self) -> &DataFlowGraph

Gets the data-flow graph associated with the function being viewed

fn current_block(&self) -> Option<Block>

Gets the current block being viewed by the cursor, if any.

fn current_inst(&self) -> Option<Inst>

Gets the current block being viewed by the cursor, if any.

fn current_block_terminator_targets(&self) -> Option<&[BlockWithParams]>

Tries to get the possible branch targets for the terminator of the current block. If there is no current block or the current block’s last instruction is not a terminator, returns None.

fn current_inst_dbg(&self) -> Option<DebugInfo>

Gets the debuginfo associated with the current instruction

fn goto_function_begin(&mut self)

Moves the position to Nothing.

fn goto_before(&mut self, block: Block)

Moves the position to Before(block).

fn goto_after(&mut self, block: Block)

Moves the position to After(block).

fn goto_first_inst(&mut self, block: Block)

Moves the position to At(block, first_inst_in_block).

fn goto_last_inst(&mut self, block: Block)

Moves the position to At(block, last_inst_in_block).

fn goto_inst(&mut self, inst: Inst)

Moves the position to At(containing, inst)

fn next_block(&mut self) -> Option<Block>

Moves the cursor to the next block in the function. If the cursor is currently not pointing to anywhere in the function, this moves it to Before(entry). If the cursor is pointing at the last block in the function, this moves it to Nothing.

fn prev_block(&mut self) -> Option<Block>

Moves the cursor to the block before the current one. If the cursor is pointing at Nothing, nothing changes. If the cursor is pointing at the first block, its moved to Nothing.

fn next_inst(&mut self) -> Option<Inst>

Moves the cursor to the next instruction in the function. If the cursor points before the block, this is the first instruction. If it points after, this does nothing. If it points at nothing, this does nothing.

fn prev_inst(&mut self) -> Option<Inst>

Moves the cursor to the previous instruction in the function. If the cursor points after the block, this is the last instruction. If it points before, this does nothing. If it points at nothing, this does nothing.

fn inst_to_result(&self, inst: Inst) -> Option<Value>

Converts an Inst into a Value that refers to the result of the instruction if possible.

Not all instructions actually yield results, those will return None

fn value_to_inst(&self, val: Value) -> Option<Inst>

Converts a Value into an Inst that yields that value, if possible.

A block param has no associated inst, that will return None.

fn is_entry_block(&self, block: Block) -> bool

Checks if a given block is the entry block to the function

fn entry_block(&self) -> Option<Block>

Gets the entry block of the function. Unless no blocks have been appended to the function, this will be Some.

fn ty(&self, value: Value) -> Type

Gets the type of a value that was previously emitted by the builder.

fn block_name(&self, block: Block) -> Str

Gets the name of a block that has been inserted into the function

fn block_params(&self, block: Block) -> &[Value]

Gets the block parameters of a given block.

fn value_def(&self, value: Value) -> ValueDef

Gets the definition of a given value

fn inst_data(&self, inst: Inst) -> &InstData

Gets the data defining a given instruction

fn inst_block(&self, inst: Inst) -> Block

Gets the block that an instruction is defined in

fn branch_info(&self, inst: Inst) -> Option<&[BlockWithParams]>

If the instruction given is a branch instruction, returns the potential branch targets.

Note that this does not care about if they are actually reachable targets, this cares about if the instruction mentions them or not. condbr false a, b would return a and b here.

fn inst_debug(&self, inst: Inst) -> DebugInfo

Returns the debug information for a given instruction

fn val_debug(&self, val: Value) -> DebugInfo

Returns the debug information for a given instruction

fn stack_slot(&self, slot: StackSlot) -> StackSlotData

Returns the information about a given stack slot