Trait sapphire::ir::InstBuilder

pub trait InstBuilder<'dfg>: Sized {

    fn dfg(&self) -> &DataFlowGraph;
    fn build(self, data: InstData, debug: DebugInfo) -> (Inst, Option<Value>);

    fn build_result(self, data: InstData, debug: DebugInfo) -> Value { ... }
    fn build_inst(self, data: InstData, debug: DebugInfo) -> Inst { ... }
    fn call(self, callee: Func, sig: Sig, args: &[Value], debug: DebugInfo) -> Inst { ... }
    fn indirectcall(
        self,
        callee: Value,
        sig: Sig,
        args: &[Value],
        debug: DebugInfo
    ) -> Inst { ... }
    fn icmp(self, cmp: ICmpOp, lhs: Value, rhs: Value, debug: DebugInfo) -> Value { ... }
    fn icmp_eq(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value { ... }
    fn icmp_ne(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value { ... }
    fn fcmp(self, cmp: FCmpOp, lhs: Value, rhs: Value, debug: DebugInfo) -> Value { ... }
    fn sel(self, cond: Value, lhs: Value, rhs: Value, debug: DebugInfo) -> Value { ... }
    fn br(self, target: BlockWithParams, debug: DebugInfo) -> Inst { ... }
    fn condbr(
        self,
        cond: Value,
        if_true: BlockWithParams,
        if_false: BlockWithParams,
        debug: DebugInfo
    ) -> Inst { ... }
    fn unreachable(self, debug: DebugInfo) -> Inst { ... }
    fn ret(self, value: Option<Value>, debug: DebugInfo) -> Inst { ... }
    fn ret_val(self, value: Value, debug: DebugInfo) -> Inst { ... }
    fn ret_void(self, debug: DebugInfo) -> Inst { ... }
    fn and(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value { ... }
    fn or(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value { ... }
    fn xor(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value { ... }
    fn shl(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value { ... }
    fn ashr(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value { ... }
    fn lshr(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value { ... }
    fn iadd(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value { ... }
    fn isub(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value { ... }
    fn imul(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value { ... }
    fn sdiv(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value { ... }
    fn udiv(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value { ... }
    fn srem(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value { ... }
    fn urem(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value { ... }
    fn fneg(self, val: Value, debug: DebugInfo) -> Value { ... }
    fn fadd(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value { ... }
    fn fsub(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value { ... }
    fn fmul(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value { ... }
    fn fdiv(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value { ... }
    fn frem(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value { ... }
    fn alloca(self, ty: Type, debug: DebugInfo) -> Value { ... }
    fn load(self, ty: Type, ptr: Value, debug: DebugInfo) -> Value { ... }
    fn store(self, val: Value, ptr: Value, debug: DebugInfo) -> Value { ... }
    fn offset(self, ty: Type, ptr: Value, offset: Value, debug: DebugInfo) -> Value { ... }
    fn extract(
        self,
        output: Type,
        agg: Value,
        index: u64,
        debug: DebugInfo
    ) -> Value { ... }
    fn insert(self, agg: Value, val: Value, index: u64, debug: DebugInfo) -> Value { ... }
    fn elemptr(self, ty: Type, ptr: Value, index: u64, debug: DebugInfo) -> Value { ... }
    fn sext(self, into: Type, from: Value, debug: DebugInfo) -> Value { ... }
    fn zext(self, into: Type, from: Value, debug: DebugInfo) -> Value { ... }
    fn trunc(self, into: Type, from: Value, debug: DebugInfo) -> Value { ... }
    fn itob(self, from: Value, debug: DebugInfo) -> Value { ... }
    fn btoi(self, into: Type, from: Value, debug: DebugInfo) -> Value { ... }
    fn sitof(self, into: Type, from: Value, debug: DebugInfo) -> Value { ... }
    fn uitof(self, into: Type, from: Value, debug: DebugInfo) -> Value { ... }
    fn ftosi(self, into: Type, from: Value, debug: DebugInfo) -> Value { ... }
    fn ftoui(self, into: Type, from: Value, debug: DebugInfo) -> Value { ... }
    fn fext(self, into: Type, from: Value, debug: DebugInfo) -> Value { ... }
    fn ftrunc(self, into: Type, from: Value, debug: DebugInfo) -> Value { ... }
    fn itop(self, from: Value, debug: DebugInfo) -> Value { ... }
    fn ptoi(self, into: Type, from: Value, debug: DebugInfo) -> Value { ... }
    fn iconst(self, into: Type, from: u64, debug: DebugInfo) -> Value { ... }
    fn fconst(self, into: Type, from: f64, debug: DebugInfo) -> Value { ... }
    fn fconst_raw(self, into: Type, raw: u64, debug: DebugInfo) -> Value { ... }
    fn bconst(self, value: bool, debug: DebugInfo) -> Value { ... }
    fn undef(self, into: Type, debug: DebugInfo) -> Value { ... }
    fn null(self, into: Type, debug: DebugInfo) -> Value { ... }

}

Helper trait that allows easy creation of instruction builders. This trait provides a variety of helper methods that build instructions and inserts them in whatever way the trait implementor defines.

This is used for the append/insert/replace builders, along with any other more situational ones scattered around the codebase.

Required Methods

fn dfg(&self) -> &DataFlowGraph

Gets the data-flow graph in use for the inserter

fn build(self, data: InstData, debug: DebugInfo) -> (Inst, Option<Value>)

“Builds” a single instruction and inserts it in whatever way the particular InstBuilder sees fit.

This returns a reference to the instruction and possibly a reference to the result of that instruction.

Provided Methods

fn build_result(self, data: InstData, debug: DebugInfo) -> Value

Builds an instruction and returns its result.

fn build_inst(self, data: InstData, debug: DebugInfo) -> Inst

Builds an instruction and returns the instruction

fn call(self, callee: Func, sig: Sig, args: &[Value], debug: DebugInfo) -> Inst

Builds a call instruction to a statically-known function.

fn indirectcall(
self,
callee: Value,
sig: Sig,
args: &[Value],
debug: DebugInfo
) -> Inst

Builds an indirectcall instruction to a statically-known function.

fn icmp(self, cmp: ICmpOp, lhs: Value, rhs: Value, debug: DebugInfo) -> Value

Builds an icmp instruction

fn icmp_eq(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value

Builds an icmp eq instruction

fn icmp_ne(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value

Builds an icmp ne instruction

fn fcmp(self, cmp: FCmpOp, lhs: Value, rhs: Value, debug: DebugInfo) -> Value

Builds an fcmp instruction

fn sel(self, cond: Value, lhs: Value, rhs: Value, debug: DebugInfo) -> Value

Builds a sel instruction

fn br(self, target: BlockWithParams, debug: DebugInfo) -> Inst

Builds a br instruction

fn condbr(
self,
cond: Value,
if_true: BlockWithParams,
if_false: BlockWithParams,
debug: DebugInfo
) -> Inst

Builds a condbr instruction

fn unreachable(self, debug: DebugInfo) -> Inst

Builds an unreachable instruction

fn ret(self, value: Option<Value>, debug: DebugInfo) -> Inst

Builds a ret instruction that possibly returns a value and possibly returns void.

fn ret_val(self, value: Value, debug: DebugInfo) -> Inst

Shorthand for Self::ret with a Some.

fn ret_void(self, debug: DebugInfo) -> Inst

Shorthand for Self::ret with a None.

fn and(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value

Builds an and instruction

fn or(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value

Builds an or instruction

fn xor(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value

Builds an xor instruction

fn shl(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value

Builds a shl instruction

fn ashr(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value

Builds an ashr instruction

fn lshr(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value

Builds a lshr instruction

fn iadd(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value

Builds an iadd instruction

fn isub(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value

Builds an isub instruction

fn imul(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value

Builds an imul instruction

fn sdiv(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value

Builds an sdiv instruction

fn udiv(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value

Builds a udiv instruction

fn srem(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value

Builds an srem instruction

fn urem(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value

Builds a urem instruction

fn fneg(self, val: Value, debug: DebugInfo) -> Value

Builds an fneg instruction

fn fadd(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value

Builds an fadd instruction

fn fsub(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value

Builds an fsub instruction

fn fmul(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value

Builds an fmul instruction

fn fdiv(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value

Builds an fdiv instruction

fn frem(self, lhs: Value, rhs: Value, debug: DebugInfo) -> Value

Builds an frem instruction

fn alloca(self, ty: Type, debug: DebugInfo) -> Value

Builds an alloca instruction

fn load(self, ty: Type, ptr: Value, debug: DebugInfo) -> Value

Builds a load instruction

fn store(self, val: Value, ptr: Value, debug: DebugInfo) -> Value

Builds a store instruction

fn offset(self, ty: Type, ptr: Value, offset: Value, debug: DebugInfo) -> Value

Builds an offset instruction

fn extract(self, output: Type, agg: Value, index: u64, debug: DebugInfo) -> Value

Builds an extract instruction that extracts a value of a given type

fn insert(self, agg: Value, val: Value, index: u64, debug: DebugInfo) -> Value

Builds an insert instruction

fn elemptr(self, ty: Type, ptr: Value, index: u64, debug: DebugInfo) -> Value

Builds an elemptr instruction

fn sext(self, into: Type, from: Value, debug: DebugInfo) -> Value

Builds a sext instruction

fn zext(self, into: Type, from: Value, debug: DebugInfo) -> Value

Builds a zext instruction

fn trunc(self, into: Type, from: Value, debug: DebugInfo) -> Value

Builds a trunc instruction

fn itob(self, from: Value, debug: DebugInfo) -> Value

Builds an itob instruction

fn btoi(self, into: Type, from: Value, debug: DebugInfo) -> Value

Builds a btoi instruction

fn sitof(self, into: Type, from: Value, debug: DebugInfo) -> Value

Builds a sitof instruction

fn uitof(self, into: Type, from: Value, debug: DebugInfo) -> Value

Builds a uitof instruction

fn ftosi(self, into: Type, from: Value, debug: DebugInfo) -> Value

Builds an ftosi instruction

fn ftoui(self, into: Type, from: Value, debug: DebugInfo) -> Value

Builds an ftoui instruction

fn fext(self, into: Type, from: Value, debug: DebugInfo) -> Value

Builds an fext instruction

fn ftrunc(self, into: Type, from: Value, debug: DebugInfo) -> Value

Builds an fext instruction

fn itop(self, from: Value, debug: DebugInfo) -> Value

Builds an itop instruction

fn ptoi(self, into: Type, from: Value, debug: DebugInfo) -> Value

Builds a ptoi instruction

fn iconst(self, into: Type, from: u64, debug: DebugInfo) -> Value

Builds an iconst instruction

fn fconst(self, into: Type, from: f64, debug: DebugInfo) -> Value

Builds an fconst instruction

fn fconst_raw(self, into: Type, raw: u64, debug: DebugInfo) -> Value

Builds an fconst instruction

fn bconst(self, value: bool, debug: DebugInfo) -> Value

Builds a bconst instruction

fn undef(self, into: Type, debug: DebugInfo) -> Value

Builds an undef instruction

fn null(self, into: Type, debug: DebugInfo) -> Value

Builds a null instruction