Trait sapphire::ir::Instruction

pub trait Instruction {
    // Required methods
    fn operands(&self) -> &[Value];
    fn result_ty(&self) -> Option<Type>;

    // Provided method
    fn has_result(&self) -> bool { ... }
}

These are the properties that any transform or analysis pass needs to be able to observe for any given instruction in any block.

Any instruction’s data can be passed as a &dyn Instruction, because every valid opcode has at least an implementation for Instruction.

Required Methods

fn operands(&self) -> &[Value]

Gets any operands that the instruction operates on.

Note that this may be an empty array, it is not safe to assume that there will be at least one operand.

fn result_ty(&self) -> Option<Type>

Gets the type of the instruction’s result after it has been evaluated.

Not all instructions will have one of these, e.g. terminators, call voids and stores do not evaluate to anything.

Provided Methods

fn has_result(&self) -> bool

Checks if the instruction yields any results at all.