Module sapphire::transforms

Defines the “transform” passes in SIR’s infrastructure.

These are the passes that can (potentially) modify SIR, and don’t actually logically yield a result.

Some of these “transforms” are not actually transformations (e.g. the verify pass is a “transform pass” even though it manipulates no IR), but most of them are. All of them logically yield no result except the IR that exists after they run.

Structs

DominatorTreeWriterPass

Wrapper pass that prints out a textual representation of a DominatorTree.

ModuleWriterAnalysis

This is a pass that writes out a textual representation of a module to a given stream.

VerifyFunctionPass

Analysis pass that wraps the verify_func function.

VerifyModulePass

Analysis pass that wraps the verify_module function.

Functions

verify_func

Verifies that a single function is a valid SIR function.

verify_module

Verifies that an entire module is valid SIR.