Module sapphire::transforms

Defines the “transform” passes in SIR’s infrastructure.

These are the passes that can (potentially) modify SIR, and don’t actually logically yield a result.

Some of these “transforms” are not actually transformations (e.g. the verify pass is a “transform pass” even though it manipulates no IR), but most of them are. All of them logically yield no result except the IR that exists after they run.

Modules

common

Common utilities that are needed between different transformations.

Structs

DeadCodeEliminationPass

Aggressive dead code elimination.

DominatorTreeWriterPass

Wrapper pass that prints out a textual representation of a DominatorTree.

GVNPass

A simple global value numbering pass.

Mem2RegPass

Promotes stack slots that only have loads and stores as uses into registers. This is effectively an SSA construction pass, it promotes memory operations into SSA values and φ nodes.

ModuleWriterPass

This is a pass that writes out a textual representation of a module to a given stream.

SCCPPass

Performs Sparse Conditional Constant Propagation.

SimplifyCFGPass

Simplifies a CFG by performing the following simplifications:

SimplifyInstPass

A “simplify all the instructions” pass. This is the algebraic simplification pass, although it will do some trivial constant folding as well. This is not a full constant propagation pass, but things like iadd 1, 1 will be simplified. Notably, it will not simplify any branches, although it may simplify their conditions.

VerifyModulePass

Analysis pass that wraps the verify_module function.

Functions

aggressive_instruction_dce

Scans over an entire function and removes dead instructions and φ nodes.

gvn

Runs a global value-numbering algorithm over func to remove redundant expressions.

mem2reg

Promotes function-local memory (that doesn’t escape) into φ nodes where possible.

verify_module

Verifies that an entire module is valid SIR.

verify_module_panic

This is verify_module, except that it writes out any errors and then panics on failure.