Trait sapphire::utility::IntoTree

pub trait IntoTree<'a, const N: usize> {
    type Node: Copy;

    // Required methods
    fn root(&'a self) -> Self::Node;
    fn children(&'a self, node: Self::Node) -> SmallVec<[Self::Node; N]>;
}

Models a type that can be traversed in a tree-like fashion. This is intended for debug APIs / passes that deal in human-readable trees.

Required Associated Types

type Node: Copy

The node type of the tree

Required Methods

fn root(&'a self) -> Self::Node

Returns the root tree node

fn children(&'a self, node: Self::Node) -> SmallVec<[Self::Node; N]>

Returns the list of children that a given node has