Trait sapphire::utility::Packable

pub trait Packable {
    fn reserved_null() -> Self;
    fn is_reserved_null(&self) -> bool;
}

Helper trait for a type that can be packed into a PackedOption.

These types need to have some null-ish value that they can reserve, that value will be used to distinguish between None and Some.

struct NonZero(i32);

impl Packable for NonZero {
    fn reserved_null() -> Self {
        NonZero(0)
    }

    fn is_reserved_null(&self) -> bool {
        self.0 == 0
    }
}

let opt = PackedOption::some(NonZero(15));

assert_eq!(opt.is_some(), true);

Required Methods

fn reserved_null() -> Self

Gets the reserved value of the type.

This value is not meant to be constructed normally in any circumstances.

fn is_reserved_null(&self) -> bool

Checks if the current object is equivalent to the constant returned by Self::reserved_null.