Trait sapphire::vm::Engine

pub trait Engine {
    // Required methods
    fn with_module(module: Module) -> Self;
    unsafe fn register_external_function<T: Any>(
        name: &str,
        func: extern "C" fn() -> T
    );
    unsafe fn run(entry: &str) -> Option<GenericForeignValue>;
    unsafe fn run_c(args: &[&str]) -> i32;
}

Abstract interface for an “engine” that can execute SIR.

How the engine decides to do that is completely irrelevant, this interface only exposes the basic operations that are necessary for a user of that engine to actually do meaningful work with it.

Required Methods

fn with_module(module: Module) -> Self

Creates an engine instance that will execute module.

unsafe fn register_external_function<T: Any>( name: &str, func: extern "C" fn() -> T )

Registers a function with a given name as the definition for a function only declared in the SIR.

Safety

This function must match the signature of the associated declaration in the SIR, this condition cannot be checked but if it is not true, the interpreter will end up calling this function with invalid arguments (and the behavior of the program will become undefined).

unsafe fn run(entry: &str) -> Option<GenericForeignValue>

Runs the entry point specified by entry. The value returned by that function (if it isn’t void) is returned.

Safety

If the SIR being executed has undefined behavior, the behavior of this call is undefined. Implementations should attempt to detect UB and abort where doing does not have a significant impact on performance, but obviously that is not possible in the general case.

unsafe fn run_c(args: &[&str]) -> i32

Runs the module as-if it was a C program, and returns the result.

This will operate as-if the C runtime is correctly initialized as needed, and the entry point is assumed to be int main(int, char**).

Safety

See the ‘Safety’ section on Self::run.